esp_idf_hal::adc
Module attenuation

Constants

DB_6

DB_2_5

DB_11

NONE

adc_atten_t_ADC_ATTEN_DB_0

<No input attenuation, ADC can measure up to approx.

adc_atten_t_ADC_ATTEN_DB_6

<The input voltage of ADC will be attenuated extending the range of measurement by about 6 dB

adc_atten_t_ADC_ATTEN_DB_2_5

<The input voltage of ADC will be attenuated extending the range of measurement by about 2.5 dB

adc_atten_t_ADC_ATTEN_DB_11

<This is deprecated, it behaves the same as ADC_ATTEN_DB_12

Type Aliases

adc_atten_t

@brief ADC attenuation parameter. Different parameters determine the range of the ADC.