esp_idf_hal::can

Enum Alert

#[repr(u32)]
pub enum Alert {

    TransmitIdle = 0,
    Success = 1,
    Received = 2,
    BelowErrorWarning = 3,
    ActiveError = 4,
    RecoveryInProgress = 5,
    BusRecovered = 6,
    ArbLost = 7,
    AboveErrorWarning = 8,
    BusError = 9,
    TransmitFailed = 10,
    ReceiveQueueFull = 11,
    ErrorPass = 12,
    BusOffline = 13,
    ReceiveFifoOverflow = 14,
    TransmitRetried = 15,
    PeripheralReset = 16,
    AlertAndLog = 17,

}

Variants

TransmitIdle = 0

No more messages to transmit

Success = 1

The previous transmission was successful

Received = 2

A frame has been received and added to the RX queue

BelowErrorWarning = 3

Both error counters have dropped below error warning limit

ActiveError = 4

TWAI controller has become error active

RecoveryInProgress = 5

TWAI controller is undergoing bus recovery

BusRecovered = 6

TWAI controller has successfully completed bus recovery

ArbLost = 7

The previous transmission lost arbitration

AboveErrorWarning = 8

One of the error counters have exceeded the error warning limit

BusError = 9

A (Bit, Stuff, CRC, Form, ACK) error has occurred on the bus

TransmitFailed = 10

The previous transmission has failed (for single shot transmission)

ReceiveQueueFull = 11

The RX queue is full causing a frame to be lost

ErrorPass = 12

TWAI controller has become error passive

BusOffline = 13

Bus-off condition occurred. TWAI controller can no longer influence bus

ReceiveFifoOverflow = 14

An RX FIFO overrun has occurred

TransmitRetried = 15

An message transmission was cancelled and retried due to an errata workaround

PeripheralReset = 16

The TWAI controller was reset

AlertAndLog = 17

In addition to alerting also Logs the event