Struct esp_idf_hal::delay::Delay

pub struct Delay(/* private fields */);

A delay provider that uses Ets for delays below a certain threshold and FreeRtos for delays equal or above the threshold.

Implementations

impl Delay

pub const fn new_default() -> Self

Create a delay with a default threshold of 1ms

pub const fn new(threshold: u32) -> Self

pub fn delay_us(&self, us: u32)

pub fn delay_ms(&self, ms: u32)