Struct esp_idf_hal::delay::Ets

pub struct Ets;

Espressif’s built-in delay provider for small delays

Use only for very small delays or else the FreeRTOS IDLE tasks might starve and the IDLE task’s watchdog will trigger.

Small delays are up to 1000 / TICK_RATE_HZ milliseconds, which is typically 10 milliseconds.

Implementations

impl Ets

pub fn delay_us(us: u32)

Pauses execution for at minimum us microseconds. The delay can be longer due to rounding and/or runtime effects.

pub fn delay_ms(ms: u32)

Pauses execution for at minimum ms milliseconds. The delay can be longer due to rounding and/or runtime effects. This delay should only be used up to 1000 / TICK_RATE_HZ milliseconds.