Struct esp_idf_hal::delay::FreeRtos

pub struct FreeRtos;

FreeRTOS-based delay provider for delays larger than 10 ms.

Delays bigger than 1000 / TICK_RATE_HZ milliseconds (typically 10 ms) used in a loop would starve the FreeRTOS IDLE tasks as they are low prio tasks and hence the IDLE task’s watchdog could trigger. This delayer avoids that by yielding to the OS during the delay.

Implementations

impl FreeRtos

pub fn delay_ms(ms: u32)

Pauses execution for at minimum ms milliseconds. The delay can be longer due to rounding and/or runtime effects.