esp_idf_hal::gpio

Struct AnyIOPin

pub struct AnyIOPin<'a> { /* private fields */ }

Implementations

impl AnyIOPin<'_>

pub unsafe fn steal(pin: PinId) -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub unsafe fn reborrow(&mut self) -> AnyIOPin<'_>

Creates a new peripheral reference with a shorter lifetime.

Use this method if you would like to keep working with the peripheral after you dropped the driver that consumes this.

Safety

You must ensure that you are not using reborrowed peripherals in drivers which are forgotten via core::mem::forget.

pub const fn none() -> Option<Self>

Return Option::None for an unconfigured instance of that pin