esp_idf_hal::gpio

Struct Gpio21

pub struct Gpio21<'a>(/* private fields */);

Implementations

impl Gpio21<'_>

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub unsafe fn reborrow(&mut self) -> Gpio21<'_>

Creates a new peripheral reference with a shorter lifetime.

Use this method if you would like to keep working with the peripheral after you dropped the driver that consumes this.

Safety

You must ensure that you are not using reborrowed peripherals in drivers which are forgotten via core::mem::forget.

impl Gpio21<'static>

pub const fn none() -> Option<Self>

Return Option::None for an unconfigured instance of that pin

impl<'a> Gpio21<'a>

pub fn degrade_input(self) -> AnyInputPin<'a>

impl<'a> Gpio21<'a>

pub fn degrade_output(self) -> AnyOutputPin<'a>

pub fn degrade_input_output(self) -> AnyIOPin<'a>