Struct esp_idf_hal::gpio::Gpio3

pub struct Gpio3<MODE> { /* private fields */ }

Implementations

impl<MODE> Gpio3<MODE> where
MODE: Send,

pub unsafe fn new() -> Gpio3<Unknown>

Safety

Care should be taken not to instantiate a pin which is already used elsewhere

pub fn into_unknown(self) -> Result<Gpio3<Unknown>, EspError>

pub fn into_disabled(self) -> Result<Gpio3<Disabled>, EspError>

pub fn into_input(self) -> Result<Gpio3<Input>, EspError>

pub unsafe fn into_subscribed(
self,
callback: impl FnMut() + Send + 'static,
interrupt_type: InterruptType
) -> Result<Gpio3<SubscribedInput>, EspError>

Safety

The callback passed to this method is executed in the context of an interrupt handler. So you should take care of what is done in it.

pub fn degrade(self) -> GpioPin<MODE>

Degrades a concrete pin (e.g. Gpio1) to a generic pin struct that can also be used with periphals.

impl<MODE> Gpio3<MODE> where
MODE: Send,

pub fn get_drive_strength(&self) -> Result<DriveStrength, EspError>

pub fn set_drive_strength(
&mut self,
strength: DriveStrength
) -> Result<(), EspError>

impl<MODE> Gpio3<MODE> where
MODE: Send,

pub fn into_input_output(self) -> Result<Gpio3<InputOutput>, EspError>

pub fn into_input_output_od(self) -> Result<Gpio3<InputOutput>, EspError>

pub fn into_output(self) -> Result<Gpio3<Output>, EspError>

pub fn into_output_od(self) -> Result<Gpio3<Output>, EspError>

impl<MODE> Gpio3<MODE> where
MODE: Send,

pub fn into_analog_atten_0db(self) -> Result<Gpio3<Atten0dB<ADC1>>, EspError>

pub fn into_analog_atten_2p5db(
self
) -> Result<Gpio3<Atten2p5dB<ADC1>>, EspError>

pub fn into_analog_atten_6db(self) -> Result<Gpio3<Atten6dB<ADC1>>, EspError>

pub fn into_analog_atten_11db(self) -> Result<Gpio3<Atten11dB<ADC1>>, EspError>