esp_idf_hal::gpio
Struct Pins

pub struct Pins {

    pub gpio0: Gpio0,
    pub gpio1: Gpio1,
    pub gpio2: Gpio2,
    pub gpio3: Gpio3,
    pub gpio4: Gpio4,
    pub gpio5: Gpio5,
    pub gpio6: Gpio6,
    pub gpio7: Gpio7,
    pub gpio8: Gpio8,
    pub gpio9: Gpio9,
    pub gpio10: Gpio10,
    pub gpio11: Gpio11,
    pub gpio12: Gpio12,
    pub gpio13: Gpio13,
    pub gpio14: Gpio14,
    pub gpio15: Gpio15,
    pub gpio16: Gpio16,
    pub gpio17: Gpio17,
    pub gpio18: Gpio18,
    pub gpio19: Gpio19,
    pub gpio20: Gpio20,
    pub gpio21: Gpio21,

}

Fields
gpio0: Gpio0gpio1: Gpio1gpio2: Gpio2gpio3: Gpio3gpio4: Gpio4gpio5: Gpio5gpio6: Gpio6gpio7: Gpio7gpio8: Gpio8gpio9: Gpio9gpio10: Gpio10gpio11: Gpio11gpio12: Gpio12gpio13: Gpio13gpio14: Gpio14gpio15: Gpio15gpio16: Gpio16gpio17: Gpio17gpio18: Gpio18gpio19: Gpio19gpio20: Gpio20gpio21: Gpio21
Implementations

impl Pins

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate the Pins structure, if it is already instantiated and used elsewhere