Trait esp_idf_hal::gpio::ADCPin

pub trait ADCPin: Pin {
    fn adc_unit(&self) -> adc_unit_t;
    fn adc_channel(&self) -> adc_channel_t;
}

A marker trait designating a pin which is capable of operating as an ADC pin, even if its current mode might be a different one

Required methods

fn adc_unit(&self) -> adc_unit_t

fn adc_channel(&self) -> adc_channel_t