esp_idf_hal::i2s::config
Enum SlotBitWidth

pub enum SlotBitWidth {
    Auto,
    Bits8,
    Bits16,
    Bits24,
    Bits32,
}

The total slot bit width in one slot.

This is not necessarily the number of data bits in one slot. A slot may have additional bits padded to fill out the slot.

Variants

Auto

Slot bit width is automatically set to the data bit width.

Bits8

Slot bit width is 8 bits.

Bits16

Slot bit width is 16 bits.

Bits24

Slot bit width is 24 bits.

Bits32

Slot bit width is 32 bits.