Struct esp_idf_hal::i2s::config::TdmClkConfig

pub struct TdmClkConfig { /* private fields */ }

TDM mode channel clock configuration.

Implementations

impl TdmClkConfig

pub fn new( sample_rate_hz: u32, clk_src: ClockSource, mclk_multiple: MclkMultiple, ) -> Self

Create a TDM clock configuration with the specified rate (in Hz), clock source, and MCLK multiple of the sample rate.

pub fn from_sample_rate_hz(rate: u32) -> Self

Create a TDM clock configuration with the specified rate in Hz. This will set the clock source to PLL_F160M, MCLK multiple to 256 times the sample rate, and MCLK to BCLK division to 8.

Note

Set the mclk_multiple to MclkMultiple::M384 when using 24-bit data width. Otherwise, the sample rate might be imprecise since the BCLK division is not an integer.

pub fn clk_src(self, clk_src: ClockSource) -> Self

Set the clock source on this TDM clock configuration.

pub fn mclk_multiple(self, mclk_multiple: MclkMultiple) -> Self

Set the MCLK multiple on this TDM clock configuration.

pub fn bclk_div(self, bclk_div: u32) -> Self

Set the MCLK to BCLK division on this TDM clock configuration.