Struct esp_idf_hal::i2s::I2sDriver

pub struct I2sDriver<'d, Dir> { /* private fields */ }

Inter-IC Sound (I2S) driver.

Implementations

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_pdm_tx<I2S: I2s>( _i2s: impl Peripheral<P = I2S> + 'd, tx_cfg: &PdmTxConfig, clk: impl Peripheral<P = impl OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd ) -> Result<Self, EspError>

Create a new pulse density modulation (PDM) mode driver for the given I2S peripheral with only the transmit channel open.

impl<'d, Dir> I2sDriver<'d, Dir>

pub fn internal_new_std<I2S: I2s>( _i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, rx: bool, tx: bool, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: Option<impl Peripheral<P = impl InputPin> + 'd>, dout: Option<impl Peripheral<P = impl OutputPin> + 'd>, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

impl<'d> I2sDriver<'d, I2sBiDir>

pub fn new_std_bidir<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with both the receive and transmit channels open.

impl<'d> I2sDriver<'d, I2sRx>

pub fn new_std_rx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with only the receive channel open.

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_std_tx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with only the transmit channel open.

impl<'d> I2sDriver<'d, I2sBiDir>

pub fn new_tdm_bidir<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with both the receive and transmit channels open.

impl<'d> I2sDriver<'d, I2sRx>

pub fn new_tdm_rx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with only the receive channel open.

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_tdm_tx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with only the transmit channel open.

impl<'d, Dir> I2sDriver<'d, Dir>

pub fn internal_new<I2S: I2s>(config: &i2s_config_t) -> Result<Self, EspError>

impl<'d, Dir> I2sDriver<'d, Dir>where Dir: I2sRxSupported,

Functions for receive channels.

pub fn rx_enable(&mut self) -> Result<(), EspError>

Enable the I2S receive channel.

Note

This can only be called when the channel is in the READY state: initialized but not yet started from a driver constructor, or disabled from the RUNNING state via rx_disable(). The channel will enter the RUNNING state if it is enabled successfully.

Enabling the channel will start I2S communications on the hardware. BCLK and WS signals will be generated if this is a controller. MCLK will be generated once initialization is finished.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the READY state.

pub fn rx_disable(&mut self) -> Result<(), EspError>

Disable the I2S receive channel.

Note

This can only be called when the channel is in the RUNNING state: the channel has been previously enabled via a call to [rx_enable()][I2sRxChannel::rx_enable]. The channel will enter the READY state if it is disabled successfully.

Disabling the channel will stop I2S communications on the hardware. BCLK and WS signals will stop being generated if this is a controller. MCLK will continue to be generated.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the RUNNING state.

pub fn read( &mut self, buffer: &mut [u8], timeout: TickType_t ) -> Result<usize, EspError>

Read data from the channel.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

pub fn read_uninit( &mut self, buffer: &mut [MaybeUninit<u8>], timeout: TickType_t ) -> Result<usize, EspError>

Read data from the channel into an uninitalized buffer.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

Safety

Upon a successful return with Ok(n_read), buffer[..n_read] will be initialized.

impl<'d, Dir> I2sDriver<'d, Dir>where Dir: I2sTxSupported,

Functions for transmit channels.

pub fn tx_enable(&mut self) -> Result<(), EspError>

Enable the I2S transmit channel.

Note

This can only be called when the channel is in the READY state: initialized but not yet started from a driver constructor, or disabled from the RUNNING state via [tx_disable()][I2sTxChannel::tx_disable]. The channel will enter the RUNNING state if it is enabled successfully.

Enabling the channel will start I2S communications on the hardware. BCLK and WS signals will be generated if this is a controller. MCLK will be generated once initialization is finished.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the READY state.

pub fn tx_disable(&mut self) -> Result<(), EspError>

Disable the I2S transmit channel.

Note

This can only be called when the channel is in the RUNNING state: the channel has been previously enabled via a call to [tx_enable()][I2sTxChannel::tx_enable]. The channel will enter the READY state if it is disabled successfully.

Disabling the channel will stop I2S communications on the hardware. BCLK and WS signals will stop being generated if this is a controller. MCLK will continue to be generated.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the RUNNING state.

pub fn write( &mut self, data: &[u8], timeout: TickType_t ) -> Result<usize, EspError>

Write data to the channel.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes sent. This may be less than the length of the data provided.

pub fn write_all( &mut self, data: &[u8], timeout: TickType_t ) -> Result<(), EspError>

Write all data to the channel.

This may be called only when the channel is in the RUNNING state.