Struct esp_idf_hal::ledc::CHANNEL3

pub struct CHANNEL3(/* private fields */);

Implementations

impl CHANNEL3

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate this peripheral instance, if it is already instantiated and used elsewhere