esp_idf_hal::ledc

Struct LedcDriver

pub struct LedcDriver<'d> { /* private fields */ }

LED Control driver

Implementations

impl<'d> LedcDriver<'d>

pub fn new<C, B>( _channel: C, timer_driver: B, pin: impl OutputPin + 'd, ) -> Result<Self, EspError>
where C: LedcChannel + 'd, B: Borrow<LedcTimerDriver<'d, C::SpeedMode>>,

Creates a new LED Control driver

pub fn config_with_pin( &mut self, pin: impl OutputPin + 'd, ) -> Result<(), EspError>

Applies LEDC configuration with a specific pin Can be used to reconfigure the LEDC driver with a different pin

pub fn get_duty(&self) -> u32

pub fn get_hpoint(&self) -> u32

pub fn get_max_duty(&self) -> u32

pub fn disable(&mut self) -> Result<(), EspError>

pub fn enable(&mut self) -> Result<(), EspError>

pub fn set_duty(&mut self, duty: u32) -> Result<(), EspError>

pub fn set_hpoint(&mut self, hpoint: u32) -> Result<(), EspError>

pub fn set_duty_with_hpoint( &mut self, duty: u32, hpoint: u32, ) -> Result<(), EspError>

pub fn channel(&self) -> ledc_channel_t

pub fn timer(&self) -> ledc_timer_t

pub fn fade_with_time( &mut self, target_duty: u32, fade_time_ms: i32, wait: bool, ) -> Result<(), EspError>

Fade the LED to a target duty cycle over a specified time

pub fn fade_with_step( &mut self, target_duty: u32, step_size: u32, step_time_ms: u32, wait: bool, ) -> Result<(), EspError>

Fade the LED to a target duty cycle using steps

pub fn fade_stop(&mut self) -> Result<(), EspError>

Stop LED fading before target duty cycle is reached