esp_idf_hal::rmt

Module encoder

Modules

simple_encoder

The SimpleEncoder provides a simpler interface than the Encoder to implement a custom RMT encoder.

Structs

BytesEncoder

An encoder that dynamically encodes a user space byte stream into RMT symbols.

BytesEncoderConfig

The configuration for the BytesEncoder.

CopyEncoder

A copy encoder copies the provided symbols from the user space into the driver layer.

EncoderWrapper

A type implementing Encoder can not be directly used by the driver, because the driver expects a different interface (see RawEncoder).

RmtChannelHandle

A handle to an RMT channel.

Enums

EncoderState

RMT encoding state

Traits

Encoder

A trait for implementing custom RMT encoders in rust.

RawEncoder

This trait represents an RMT encoder that is used to encode data for transmission.

Functions

into_raw

A helper function to convert a custom RMT encoder into a raw encoder that can be used by the RMT driver.