esp_idf_hal::rmt::encoder

Struct BytesEncoder

pub struct BytesEncoder { /* private fields */ }

An encoder that dynamically encodes a user space byte stream into RMT symbols.

It is usually used to encode dynamic data, e.g., the address and command fields in the IR protocol.

Implementations

impl BytesEncoder

pub fn new() -> Result<Self, EspError>

Constructs a new bytes encoder with default configuration.

pub fn with_config(config: &BytesEncoderConfig) -> Result<Self, EspError>

Constructs a new bytes encoder with the provided configuration.