esp_idf_hal::rmt::encoder

Trait RawEncoder

pub trait RawEncoder {
    type Item;

    // Required method
    fn handle(&mut self) -> &mut rmt_encoder_t;

    // Provided method
    fn reset(&mut self) -> Result<(), EspError> { ... }
}

This trait represents an RMT encoder that is used to encode data for transmission.

Required Associated Types

type Item

The type of input data the encoder can encode.

Required Methods

fn handle(&mut self) -> &mut rmt_encoder_t

Returns a mutable reference to the underlying rmt_encoder_t.

The functions of the rmt_encoder_t will be called by the RMT driver.

Provided Methods

fn reset(&mut self) -> Result<(), EspError>

Reset the encoder state.

Errors

If resetting the encoder fails, it should return an EspError with the code ESP_FAIL.