esp_idf_hal::rmt
Struct CHANNEL0

pub struct CHANNEL0(/* private fields */);

Implementations

impl CHANNEL0

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate this peripheral instance, if it is already instantiated and used elsewhere