esp_idf_hal::rmt
Struct Pulse

pub struct Pulse {
    pub ticks: PulseTicks,
    pub pin_state: PinState,
}

A Pulse contains a pin state and a tick count, used in creating a Signal.

The real time duration of a tick depends on the [Config::clock_divider] setting.

You can create a Pulse with a [Duration] by using Pulse::new_with_duration.

Example

let pulse = Pulse::new(PinState::High, PulseTicks::new(32));

Fields
ticks: PulseTickspin_state: PinState
Implementations

impl Pulse

pub const fn zero() -> Self

pub const fn new(pin_state: PinState, ticks: PulseTicks) -> Self

pub fn new_with_duration( ticks_hz: Hertz, pin_state: PinState, duration: &Duration, ) -> Result<Self, EspError>

Create a Pulse using a [Duration].

The ticks frequency, which depends on the clock divider set on the channel within a [Transmit]. To get the frequency for the ticks_hz argument, use [Transmit::counter_clock()].

Example

let tx = Transmit::new(led, channel, &config)?;
let ticks_hz = tx.counter_clock()?;
let pulse = Pulse::new_with_duration(ticks_hz, PinState::High, Duration::from_nanos(500))?;