esp_idf_hal::rmt

Struct Symbol

pub struct Symbol(/* private fields */);

A Symbol constists of two Pulses, where each pulse defines a level of the pin (high or low) and a duration (PulseTicks).

The Pulses can be the same level (e.g., both high) and must not necessarily be different levels.

This is just a newtype over the IDF’s rmt_item32_t or rmt_symbol_word_t type.

Implementations

impl Symbol

pub fn new(level0: Pulse, level1: Pulse) -> Self

Create a symbol from a pair of half-cycles.

pub fn new_with( resolution: Hertz, level0: PinState, duration0: Duration, level1: PinState, duration1: Duration, ) -> Result<Self, EspError>

Constructs a symbol from the given levels and durations.

This is a convenience function that combines Pulse::new_with_duration and Symbol::new.

pub fn new_half_split( resolution: Hertz, level0: PinState, level1: PinState, duration: Duration, ) -> Result<Self, EspError>

Constructs a new symbol where the duration is split evenly between the two levels.

It is guaranteed that the combined duration of the two levels adds up to the given duration. If the duration is odd, the first level might be a bit shorter than the second level.

pub fn repeat_for( &self, resolution: Hertz, duration: Duration, ) -> impl Iterator<Item = Self>

Repeats the symbol to have the returned sequence of symbols last for exactly the given duration.

There is an upper limit for how long a single symbol can be (PulseTicks::max). To create a symbol that lasts longer than that, it is split into multiple symbols. The returned iterator yields as many symbols as necessary to reach the desired duration.

If the given duration is not a multiple of the symbol duration, the last symbol will be adjusted to occupy the remaining time.

Panics

If self has a duration of zero.

pub fn level0(&self) -> Pulse

Returns the first half-cycle (pulse) of this symbol.

pub fn level1(&self) -> Pulse

Returns the second half-cycle (pulse) of this symbol.

pub fn duration(&self, resolution: Hertz) -> Duration

Returns the combined duration of both half-cycles of this symbol.

pub fn update(&mut self, level0: Pulse, level1: Pulse)

Mutate this symbol to store a different pair of half-cycles.