esp_idf_hal::rmt

Struct TxQueue

pub struct TxQueue<'c, 'd, E: Encoder> { /* private fields */ }

A queue to efficiently transmit multiple signals, reusing encoders when possible.

The driver has an internal queue from which it will copy the signals into the peripheral memory. The signal and encoder must be valid until the transmission is done, which can not be guaranteed when directly using the queue through TxChannelDriver::start_send.

This type provides a safe interface to fill up the queue, reusing encoders when possible. To do this, it keeps an allocated buffer for each encoder, to which the signal will be copied before starting the transmission. This ensures that both the signal and encoder are valid until the transmission is done.

Drop behavior

When the TxQueue is dropped, it will wait for all transmissions to finish. This ensures that the internal buffers are not dropped while they are still in use by the peripheral.

Implementations

impl<'c, 'd, E: Encoder> TxQueue<'c, 'd, E>

pub fn channel(&mut self) -> &mut TxChannelDriver<'d>

Returns a mutable reference to the channel this queue is using.

impl<'c, 'd, E: Encoder> TxQueue<'c, 'd, E>

pub fn push( &mut self, signal: &[E::Item], config: &TransmitConfig, ) -> Result<(), EspError>
where E::Item: Clone,

Pushes a signal onto the transmission queue.

The signal will be cloned into an internal buffer to ensure that it is valid for the entire transmission. These buffers will be reused for future transmissions after the current transmission is done.

Blocking behavior

If the queue is full (i.e. all encoders are busy), this function will behave differently depending on whether TransmitConfig::queue_non_blocking is set or not:

If queue_non_blocking is false, it will block until one encoder is available.

If queue_non_blocking is true, it will return an error with code ESP_ERR_TIMEOUT. There will be no state change, so the function can be called again immediately.

If this queue has more encoders available than the channel queue size and queue_non_blocking is set, it will copy the signal into the internal buffer, but might fail (with a timeout) to start the transmission if the channel queue is full.

Panics

If the signal is empty.