esp_idf_hal::rmt
Struct VariableLengthSignal

pub struct VariableLengthSignal { /* private fields */ }

Vec heap based storage for an RMT signal.

Use this for when you donโt know the final size of your signal data.

Example

let mut signal = VariableLengthSignal::new();
signal.push(Pulse::new(PinState::High, PulseTicks::new(10)));
signal.push(Pulse::new(PinState::Low, PulseTicks::new(9)));

Implementations

impl VariableLengthSignal

pub const fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

Create a new VariableLengthSignal with a a given capacity. This is more efficent than not specifying the capacity with new( ) as the memory manager only needs to allocate the underlying array once.

capacity is the number of Pulses which can be pushes before reallocating

pub fn push<'p, I>(&mut self, pulses: I) -> Result<(), EspError>
where I: IntoIterator<Item = &'p Pulse>,

Add Pulses to the end of the signal.

pub fn clear(&mut self)

Delete all pulses.