esp_idf_hal::sd

Struct SdCardDriver

pub struct SdCardDriver<T> { /* private fields */ }

A high-level SD-Card driver.

This driver is used to interface with an SD-Card by wrapping one of the two SD Host drivers:

SD-SPI Host driver (SdSpiHostDriver)

SD-MMC Host driver (SdMmcHostDriver) - on MCUs that do have an SD-MMC peripheral (ESP32, ESP32S3 and ESP32P4)

The interface allows reading, writing and erasing sectors, as well as reading and writing arbitrary-length bytes.

Currently, all interaction with the SD-Card driver is via the native, unsafe sys::sdmmc_* functions.

Implementations

impl<T> SdCardDriver<T>

pub fn card(&self) -> &sdmmc_card_t

Get a reference to the SD-Card native structure.

impl<'d, T> SdCardDriver<SdSpiHostDriver<'d, T>>
where T: Borrow<SpiDriver<'d>>,

pub fn new_spi( host: SdSpiHostDriver<'d, T>, configuration: &Configuration, ) -> Result<Self, EspError>

Create a new SD-Card driver using the SD-SPI host driver instantiated with one of the SPI peripherals