Module esp_idf_hal::serial

UART peripheral control

Controls UART peripherals (UART0, UART1, UART2). Notice that UART0 is typically already used for loading firmware and logging. Therefore use UART1 and UART2 in your application. Any pin can be used for rx and tx.

Example

Create a serial peripheral and write to serial port.

use std::fmt::Write;
use esp_idf_hal::prelude::*;
use esp_idf_hal::serial;

let peripherals = Peripherals::take().unwrap();
let pins = peripherals.pins;

let config = serial::config::Config::default().baudrate(Hertz(115_200));

let mut serial: serial::Serial<serial::UART1, _, _> = serial::Serial::new(
    peripherals.uart1,
    serial::Pins {
        tx: pins.gpio1,
        rx: pins.gpio3,
        cts: None,
        rts: None,
    },
    config
).unwrap();

for i in 0..10 {
    writeln!(serial, "{:}", format!("count {:}", i)).unwrap();
}

TODO

Add all extra features esp32 supports (eg rs485, etc. etc.)

Free APB lock when TX is idle (and no RX used)

Address errata 3.17: UART fifo_cnt is inconsistent with FIFO pointer

Modules

config

UART configuration

Structs

Pins

Pins used by the UART interface

Rx

Serial receiver

Serial

Serial abstraction

SerialError

Tx

Serial transmitter

UART0

UART1

Traits

Uart