Struct esp_idf_hal::serial::Serial

pub struct Serial<UART: Uart, TX: OutputPin, RX: InputPin, CTS: InputPin = Gpio1<Input>, RTS: OutputPin = Gpio2<Output>> { /* private fields */ }

Serial abstraction

Implementations

impl<UART: Uart, TX: OutputPin, RX: InputPin, CTS: InputPin, RTS: OutputPin> Serial<UART, TX, RX, CTS, RTS>

pub fn new(
uart: UART,
pins: Pins<TX, RX, CTS, RTS>,
config: Config
) -> Result<Self, EspError>

Create a new serial driver

pub fn change_stop_bits(
&mut self,
stop_bits: StopBits
) -> Result<&mut Self, EspError>

Change the number of stop bits

pub fn stop_bits(&self) -> Result<StopBits, EspError>

Retruns the current number of stop bits

pub fn change_data_bits(
&mut self,
data_bits: DataBits
) -> Result<&mut Self, EspError>

Change the number of data bits

pub fn data_bits(&self) -> Result<DataBits, EspError>

Return the current number of data bits

pub fn change_parity(&mut self, parity: Parity) -> Result<&mut Self, EspError>

Change the type of parity checking

pub fn parity(&self) -> Result<Parity, EspError>

Returns the current type of parity checking

pub fn change_baudrate<T: Into<Hertz> + Copy>(
&mut self,
baudrate: T
) -> Result<&mut Self, EspError>

Change the baudrate.

Will automatically select the clock source. When possible the reference clock (1MHz) will be used, because this is constant when the clock source/frequency changes. However if one of the clock frequencies is below 10MHz or if the baudrate is above the reference clock or if the baudrate cannot be set within 1.5% then use the APB clock.

pub fn baudrate(&self) -> Result<Hertz, EspError>

Returns the current baudrate

pub fn split(self) -> (Tx<UART>, Rx<UART>)

Split the serial driver in separate TX and RX drivers

pub fn release(self) -> Result<(UART, Pins<TX, RX, CTS, RTS>), EspError>

Release the UART and GPIO resources