Struct esp_idf_hal::spi::config::Config

pub struct Config {
    pub baudrate: Hertz,
    pub data_mode: Mode,
    pub write_only: bool,
    pub duplex: Duplex,
    pub bit_order: BitOrder,
    pub cs_active_high: bool,
    pub input_delay_ns: i32,
    pub polling: bool,
    pub allow_pre_post_delays: bool,
    pub queue_size: usize,
}

SPI Device configuration

Fields
baudrate: Hertzdata_mode: Modewrite_only: bool

This property can be set to configure a SPI Device for being write only Thus the flag SPI_DEVICE_NO_DUMMY will be passed on initialization and it will unlock the possibility of using 80Mhz as the bus freq See https://docs.espressif.com/projects/esp-idf/en/latest/esp32/api-reference/peripherals/spi_master.html#timing-considerations

duplex: Duplexbit_order: BitOrdercs_active_high: boolinput_delay_ns: i32polling: boolallow_pre_post_delays: boolqueue_size: usize
Implementations

impl Config

pub fn new() -> Self

pub fn baudrate(self, baudrate: Hertz) -> Self

pub fn data_mode(self, data_mode: Mode) -> Self

pub fn write_only(self, write_only: bool) -> Self

pub fn duplex(self, duplex: Duplex) -> Self

pub fn bit_order(self, bit_order: BitOrder) -> Self

pub fn cs_active_high(self) -> Self

pub fn input_delay_ns(self, input_delay_ns: i32) -> Self

pub fn polling(self, polling: bool) -> Self

pub fn allow_pre_post_delays(self, allow_pre_post_delays: bool) -> Self

pub fn queue_size(self, queue_size: usize) -> Self