esp_idf_hal::spi
Struct SpiBusDriver

pub struct SpiBusDriver<'d, T>
where
    T: BorrowMut<SpiDriver<'d>>,
{ /* private fields */ }

Implementations

impl<'d, T> SpiBusDriver<'d, T>
where T: BorrowMut<SpiDriver<'d>>,

pub fn new(driver: T, config: &Config) -> Result<Self, EspError>

pub fn read(&mut self, words: &mut [u8]) -> Result<(), EspError>

pub async fn read_async(&mut self, words: &mut [u8]) -> Result<(), EspError>

pub fn write(&mut self, words: &[u8]) -> Result<(), EspError>

pub async fn write_async(&mut self, words: &[u8]) -> Result<(), EspError>

pub fn transfer( &mut self, read: &mut [u8], write: &[u8], ) -> Result<(), EspError>

pub async fn transfer_async( &mut self, read: &mut [u8], write: &[u8], ) -> Result<(), EspError>

pub fn transfer_in_place(&mut self, words: &mut [u8]) -> Result<(), EspError>

pub async fn transfer_in_place_async( &mut self, words: &mut [u8], ) -> Result<(), EspError>

pub fn flush(&mut self) -> Result<(), EspError>

pub fn operation(&mut self, operation: Operation<'_>) -> Result<(), EspError>

Run the provided Operation on the bus.

Only Operations that result in a transfer are supported. For example, passing an Operation::DelayNs will return an error.

pub async fn operation_async( &mut self, operation: Operation<'_>, ) -> Result<(), EspError>

Run the provided Operation on the bus.

Only Operations that result in a transfer are supported. For example, passing an Operation::DelayNs will return an error.