esp_idf_hal::spi
Struct SpiSoftCsDeviceDriver

pub struct SpiSoftCsDeviceDriver<'d, DEVICE, DRIVER> { /* private fields */ }

Implementations

impl<'d, DEVICE, DRIVER> SpiSoftCsDeviceDriver<'d, DEVICE, DRIVER>
where DEVICE: Borrow<SpiSharedDeviceDriver<'d, DRIVER>>, DRIVER: Borrow<SpiDriver<'d>> + 'd,

pub fn new( shared_device: DEVICE, cs: impl Peripheral<P = impl OutputPin> + 'd, cs_level: Level, ) -> Result<Self, EspError>

pub fn cs_pre_delay_us(&mut self, delay_us: u32) -> &mut Self

Add an aditional Amount of SPI bit-cycles the cs should be activated before the transmission (0-16). This only works on half-duplex transactions.

pub fn cs_post_delay_us(&mut self, delay_us: u32) -> &mut Self

Add an aditional delay of x in uSeconds after transaction between last clk out and chip select

pub fn transaction( &mut self, operations: &mut [Operation<'_>], ) -> Result<(), EspError>

pub async fn transaction_async( &mut self, operations: &mut [Operation<'_>], ) -> Result<(), EspError>

pub fn read(&mut self, read: &mut [u8]) -> Result<(), EspError>

pub async fn read_async(&mut self, read: &mut [u8]) -> Result<(), EspError>

pub fn write(&mut self, write: &[u8]) -> Result<(), EspError>

pub async fn write_async(&mut self, write: &[u8]) -> Result<(), EspError>

pub fn transfer_in_place(&mut self, buf: &mut [u8]) -> Result<(), EspError>

pub async fn transfer_in_place_async( &mut self, buf: &mut [u8], ) -> Result<(), EspError>

pub fn transfer( &mut self, read: &mut [u8], write: &[u8], ) -> Result<(), EspError>

pub async fn transfer_async( &mut self, read: &mut [u8], write: &[u8], ) -> Result<(), EspError>