esp_idf_hal::sys
Function esp_adc_cal_check_efuse

pub unsafe extern "C" fn esp_adc_cal_check_efuse(
    value_type: u32,
) -> i32

@brief Checks if ADC calibration values are burned into eFuse

This function checks if ADC reference voltage or Two Point values have been burned to the eFuse of the current ESP32

@param value_type Type of calibration value (ESP_ADC_CAL_VAL_EFUSE_VREF or ESP_ADC_CAL_VAL_EFUSE_TP) @note in ESP32S2, only ESP_ADC_CAL_VAL_EFUSE_TP is supported. Some old ESP32S2s do not support this, either. In which case you have to calibrate it manually, possibly by performing your own two-point calibration on the chip.

@return - ESP_OK: The calibration mode is supported in eFuse - ESP_ERR_NOT_SUPPORTED: Error, eFuse values are not burned - ESP_ERR_INVALID_ARG: Error, invalid argument (ESP_ADC_CAL_VAL_DEFAULT_VREF)