esp_idf_hal::sys
Function esp_ble_gattc_get_service

pub unsafe extern "C" fn esp_ble_gattc_get_service(
    gattc_if: u8,
    conn_id: u16,
    svc_uuid: *mut esp_bt_uuid_t,
    result: *mut esp_gattc_service_elem_t,
    count: *mut u16,
    offset: u16,
) -> u32

@brief Find all the service with the given service uuid in the gattc cache, if the svc_uuid is NULL, find all the service. Note: It just get service from local cache, won’t get from remote devices. If want to get it from remote device, need to used the esp_ble_gattc_cache_refresh, then call esp_ble_gattc_get_service again.

@param[in] gattc_if: Gatt client access interface. @param[in] conn_id: connection ID which identify the server. @param[in] svc_uuid: the pointer to the service uuid. @param[out] result: The pointer to the service which has been found in the gattc cache. @param[inout] count: input the number of service want to find, it will output the number of service has been found in the gattc cache with the given service uuid. @param[in] offset: Offset of the service position to get.

@return - ESP_OK: success - other: failed