esp_idf_hal::sys

Function esp_https_ota_finish

pub unsafe extern "C" fn esp_https_ota_finish(
    https_ota_handle: *mut c_void,
) -> i32

@brief Clean-up HTTPS OTA Firmware upgrade and close HTTPS connection

This function closes the HTTP connection and frees the ESP HTTPS OTA context. This function switches the boot partition to the OTA partition containing the new firmware image.

@note If this API returns successfully, esp_restart() must be called to boot from the new firmware image esp_https_ota_finish should not be called after calling esp_https_ota_abort

@param[in] https_ota_handle pointer to esp_https_ota_handle_t structure

@return

ESP_OK: Clean-up successful

ESP_ERR_INVALID_STATE

ESP_ERR_INVALID_ARG: Invalid argument

ESP_ERR_OTA_VALIDATE_FAILED: Invalid app image