esp_idf_hal::sys

Function esp_netif_action_remove_ip6_address

pub unsafe extern "C" fn esp_netif_action_remove_ip6_address(
    esp_netif: *mut c_void,
    base: *const u8,
    event_id: i32,
    data: *mut c_void,
)

@brief Default building block for network interface action upon IPv6 address removed by the underlying stack

@note This API can be directly used as event handler

@param[in] esp_netif Handle to esp-netif instance @param base The base type of the event @param event_id The specific ID of the event @param data Optional data associated with the event