esp_idf_hal::sys

Function esp_netif_transmit_wrap

pub unsafe extern "C" fn esp_netif_transmit_wrap(
    esp_netif: *mut esp_netif_obj,
    data: *mut c_void,
    len: usize,
    netstack_buf: *mut c_void,
) -> i32

@brief Outputs packets from the TCP/IP stack to the media to be transmitted

This function gets called from network stack to output packets to IO driver.

@param[in] esp_netif Handle to esp-netif instance @param[in] data Data to be transmitted @param[in] len Length of the data frame @param[in] netstack_buf net stack buffer

@return ESP_OK on success, an error passed from the I/O driver otherwise