esp_idf_hal::sys

Function esp_tls_conn_new_sync

pub unsafe extern "C" fn esp_tls_conn_new_sync(
    hostname: *const u8,
    hostlen: i32,
    port: i32,
    cfg: *const esp_tls_cfg,
    tls: *mut esp_tls,
) -> i32

@brief Create a new blocking TLS/SSL connection

This function establishes a TLS/SSL connection with the specified host in blocking manner.

@param[in] hostname Hostname of the host. @param[in] hostlen Length of hostname. @param[in] port Port number of the host. @param[in] cfg TLS configuration as esp_tls_cfg_t. For a TLS connection, pass a pointer to a esp_tls_cfg_t. For a plain TCP connection, pass a pointer to a esp_tls_cfg_t with is_plain_tcp set to true. At a minimum, this pointer should be not NULL and the structure should be zero-initialized @param[in] tls Pointer to esp-tls as esp-tls handle.

@return - -1 If connection establishment fails. - 1 If connection establishment is successful. - 0 If connection state is in progress.