esp_idf_hal::sys

Function esp_vfs_register_fd

pub unsafe extern "C" fn esp_vfs_register_fd(
    vfs_id: i32,
    fd: *mut i32,
) -> i32

Special function for registering another file descriptor for a VFS registered by esp_vfs_register_with_id. This function should only be used to register permanent file descriptors (socket fd) that are not removed after being closed.

@param vfs_id VFS identificator returned by esp_vfs_register_with_id. @param fd The registered file descriptor will be written to this address.

@return ESP_OK if the registration is successful, ESP_ERR_NO_MEM if too many file descriptors are registered, ESP_ERR_INVALID_ARG if the arguments are incorrect.