esp_idf_hal::sys

Function esp_wifi_ftm_initiate_session

pub unsafe extern "C" fn esp_wifi_ftm_initiate_session(
    cfg: *mut wifi_ftm_initiator_cfg_t,
) -> i32

@brief Start an FTM Initiator session by sending FTM request If successful, event WIFI_EVENT_FTM_REPORT is generated with the result of the FTM procedure

@attention 1. Use this API only in Station mode. @attention 2. If FTM is initiated on a different channel than Station is connected in or internal SoftAP is started in, FTM defaults to a single burst in ASAP mode.

@param cfg FTM Initiator session configuration

@return

ESP_OK: succeed

others: failed