Function esp_idf_hal::sys::esp_wifi_internal_set_fix_rate

pub unsafe extern "C" fn esp_wifi_internal_set_fix_rate(
    ifx: u32,
    en: bool,
    rate: u32
) -> i32

@brief enable or disable transmitting WiFi MAC frame with fixed rate

@attention 1. If fixed rate is enabled, both management and data frame are transmitted with fixed rate @attention 2. Make sure that the receiver is able to receive the frame with the fixed rate if you want the frame to be received @attention 3. Not support to set fix rate for espnow and 80211_tx

@param ifx : wifi interface @param en : false - disable, true - enable @param rate : PHY rate

@return

ERR_OK : succeed

ESP_ERR_WIFI_NOT_INIT: WiFi is not initialized by esp_wifi_init

ESP_ERR_WIFI_NOT_STARTED: WiFi was not started by esp_wifi_start

ESP_ERR_WIFI_IF : invalid WiFi interface

ESP_ERR_INVALID_ARG : invalid rate

ESP_ERR_NOT_SUPPORTED : do not support to set fixed rate if TX AMPDU is enabled