Function esp_idf_hal::sys::esp_wifi_set_rssi_threshold

pub unsafe extern "C" fn esp_wifi_set_rssi_threshold(rssi: i32) -> i32

@brief Set RSSI threshold, if average rssi gets lower than threshold, WiFi task will post event WIFI_EVENT_STA_BSS_RSSI_LOW.

@attention If the user wants to receive another WIFI_EVENT_STA_BSS_RSSI_LOW event after receiving one, this API needs to be called again with an updated/same RSSI threshold.

@param rssi threshold value in dbm between -100 to 0

@return

ESP_OK: succeed

ESP_ERR_WIFI_NOT_INIT: WiFi is not initialized by esp_wifi_init

ESP_ERR_INVALID_ARG: invalid argument