esp_idf_hal::sys
Function i2s_channel_reconfig_tdm_clock

pub unsafe extern "C" fn i2s_channel_reconfig_tdm_clock(
    handle: *mut i2s_channel_obj_t,
    clk_cfg: *const i2s_tdm_clk_config_t,
) -> i32

@brief Reconfigure the I2S clock for TDM mode @note Only allowed to be called when the channel state is READY, i.e., channel has been initialized, but not started this function won’t change the state. i2s_channel_disable should be called before calling this function if I2S has started. @note The input channel handle has to be initialized to TDM mode, i.e., i2s_channel_init_tdm_mode has been called before reconfiguring

@param[in] handle I2S channel handler @param[in] clk_cfg Standard mode clock configuration, can be generated by I2S_TDM_CLK_DEFAULT_CONFIG @return - ESP_OK Set clock successfully - ESP_ERR_INVALID_ARG NULL pointer, invalid configuration or not TDM mode - ESP_ERR_INVALID_STATE This channel is not initialized or not stopped