esp_idf_hal::sys
Function i2s_set_pin

pub unsafe extern "C" fn i2s_set_pin(
    i2s_num: u32,
    pin: *const i2s_pin_config_t,
) -> i32

@brief Set I2S pin number

@note The I2S peripheral output signals can be connected to multiple GPIO pads. However, the I2S peripheral input signal can only be connected to one GPIO pad.

@param i2s_num I2S port number

@param pin I2S Pin structure, or NULL to set 2-channel 8-bit internal DAC pin configuration (GPIO25 & GPIO26)

Inside the pin configuration structure, set I2S_PIN_NO_CHANGE for any pin where the current configuration should not be changed.

@note if *pin is set as NULL, this function will initialize both of the built-in DAC channels by default. if you don’t want this to happen and you want to initialize only one of the DAC channels, you can call i2s_set_dac_mode instead.

@return - ESP_OK Success - ESP_ERR_INVALID_ARG Parameter error - ESP_FAIL IO error