esp_idf_hal::sys

Function mbedtls_asn1_sequence_free

pub unsafe extern "C" fn mbedtls_asn1_sequence_free(
    seq: *mut mbedtls_asn1_sequence,
)

\brief Free a heap-allocated linked list presentation of an ASN.1 sequence, including the first element.

There are two common ways to manage the memory used for the representation of a parsed ASN.1 sequence:

Allocate a head node mbedtls_asn1_sequence *head with mbedtls_calloc(). Pass this node as the cur argument to mbedtls_asn1_get_sequence_of(). When you have finished processing the sequence, call mbedtls_asn1_sequence_free() on head.

Allocate a head node mbedtls_asn1_sequence *head in any manner, for example on the stack. Make sure that head->next == NULL. Pass head as the cur argument to mbedtls_asn1_get_sequence_of(). When you have finished processing the sequence, call mbedtls_asn1_sequence_free() on head->cur, then free head itself in the appropriate manner.

\param seq The address of the first sequence component. This may be \c NULL, in which case this functions returns immediately.