esp_idf_hal::sys
Function mbedtls_cipher_cmac_update

pub unsafe extern "C" fn mbedtls_cipher_cmac_update(
    ctx: *mut mbedtls_cipher_context_t,
    input: *const u8,
    ilen: usize,
) -> i32

\brief This function feeds an input buffer into an ongoing CMAC computation.

                 The CMAC computation must have previously been started
                 by calling mbedtls_cipher_cmac_starts() or
                 mbedtls_cipher_cmac_reset().

                 Call this function as many times as needed to input the
                 data to be authenticated.
                 Once all of the required data has been input,
                 call mbedtls_cipher_cmac_finish() to obtain the result
                 of the CMAC operation.

\param ctx The cipher context used for the CMAC operation. \param input The buffer holding the input data. \param ilen The length of the input data.

\return \c 0 on success. \return #MBEDTLS_ERR_MD_BAD_INPUT_DATA if parameter verification fails.