Function esp_idf_hal::sys::mbedtls_ecp_tls_write_point

pub unsafe extern "C" fn mbedtls_ecp_tls_write_point(
    grp: *const mbedtls_ecp_group,
    pt: *const mbedtls_ecp_point,
    format: i32,
    olen: *mut usize,
    buf: *mut u8,
    blen: usize,
) -> i32

\brief This function exports a point as a TLS ECPoint record defined in RFC 4492, Section 5.4.

\param grp The ECP group to use. This must be initialized and have group parameters set, for example through mbedtls_ecp_group_load(). \param pt The point to be exported. This must be initialized. \param format The point format to use. This must be either #MBEDTLS_ECP_PF_COMPRESSED or #MBEDTLS_ECP_PF_UNCOMPRESSED. \param olen The address at which to store the length in Bytes of the data written. \param buf The target buffer. This must be a writable buffer of length \p blen Bytes. \param blen The length of the target buffer \p buf in Bytes.

\return \c 0 on success. \return #MBEDTLS_ERR_ECP_BAD_INPUT_DATA if the input is invalid. \return #MBEDTLS_ERR_ECP_BUFFER_TOO_SMALL if the target buffer is too small to hold the exported point. \return Another negative error code on other kinds of failure.