esp_idf_hal::sys
Function mcpwm_operator_apply_carrier

pub unsafe extern "C" fn mcpwm_operator_apply_carrier(
    oper: *mut mcpwm_oper_t,
    config: *const mcpwm_carrier_config_t,
) -> i32

@brief Apply carrier feature for MCPWM operator

@param[in] oper MCPWM operator, allocated by mcpwm_new_operator() @param[in] config MCPWM carrier specific configuration @return - ESP_OK: Set carrier for operator successfully - ESP_ERR_INVALID_ARG: Set carrier for operator failed because of invalid argument - ESP_FAIL: Set carrier for operator failed because of other error