esp_idf_hal::sys

Function nvs_flash_register_security_scheme

pub unsafe extern "C" fn nvs_flash_register_security_scheme(
    scheme_cfg: *mut nvs_sec_scheme_t,
) -> i32

@brief Registers the given security scheme for NVS encryption The scheme registered with sec_scheme_id by this API be used as the default security scheme for the “nvs” partition. Users will have to call this API explicitly in their application.

@param[in] scheme_cfg Pointer to the security scheme configuration structure that the user (or the nvs_key_provider) wants to register.

@return - ESP_OK, if security scheme registration succeeds; - ESP_ERR_INVALID_ARG, if scheme_cfg is NULL; - ESP_FAIL, if security scheme registration fails