esp_idf_hal::sys
Function otLinkFilterSetDefaultRssIn

pub unsafe extern "C" fn otLinkFilterSetDefaultRssIn(
    aInstance: *mut otInstance,
    aRss: i8,
)

Sets the default received signal strength (in dBm) on MAC Filter.

Is available when OPENTHREAD_CONFIG_MAC_FILTER_ENABLE configuration is enabled.

The default RSS value is used for all received frames from addresses for which there is no explicit RSS-IN entry in the Filter list (added using otLinkFilterAddRssIn()).

@param[in] aInstance A pointer to an OpenThread instance. @param[in] aRss The default received signal strength (in dBm) to set.