esp_idf_hal::sys
Function otThreadDiscover

pub unsafe extern "C" fn otThreadDiscover(
    aInstance: *mut otInstance,
    aScanChannels: u32,
    aPanId: u16,
    aJoiner: bool,
    aEnableEui64Filtering: bool,
    aCallback: Option<unsafe extern "C" fn(_: *mut otActiveScanResult, _: *mut c_void)>,
    aCallbackContext: *mut c_void,
) -> u32

Starts a Thread Discovery scan.

@note A successful call to this function enables the rx-on-when-idle mode for the entire scan procedure.

@param[in] aInstance A pointer to an OpenThread instance. @param[in] aScanChannels A bit vector indicating which channels to scan (e.g. OT_CHANNEL_11_MASK). @param[in] aPanId The PAN ID filter (set to Broadcast PAN to disable filter). @param[in] aJoiner Value of the Joiner Flag in the Discovery Request TLV. @param[in] aEnableEui64Filtering TRUE to filter responses on EUI-64, FALSE otherwise. @param[in] aCallback A pointer to a function called on receiving an MLE Discovery Response or scan completes. @param[in] aCallbackContext A pointer to application-specific context.

@retval OT_ERROR_NONE Successfully started a Thread Discovery Scan. @retval OT_ERROR_INVALID_STATE The IPv6 interface is not enabled (netif is not up). @retval OT_ERROR_NO_BUFS Could not allocate message for Discovery Request. @retval OT_ERROR_BUSY Thread Discovery Scan is already in progress.