esp_idf_hal::sys

Function twai_clear_transmit_queue_v2

pub unsafe extern "C" fn twai_clear_transmit_queue_v2(
    handle: *mut twai_obj_t,
) -> i32

@brief Clear the transmit queue of a given TWAI driver handle

@note This is an advanced version of twai_clear_transmit_queue that can clear the transmit queue of a given TWAI driver handle. Please refer to the documentation of twai_clear_transmit_queue for more details.

@param[in] handle TWAI driver handle returned by twai_driver_install_v2

@return - ESP_OK: Transmit queue cleared - ESP_ERR_INVALID_STATE: TWAI driver is not installed or TX queue is disabled