esp_idf_hal::sys

Function uxTaskPriorityGet

pub unsafe extern "C" fn uxTaskPriorityGet(
    xTask: *mut tskTaskControlBlock,
) -> u32

INCLUDE_uxTaskPriorityGet must be defined as 1 for this function to be available. See the configuration section for more information.

Obtain the priority of any task.

@param xTask Handle of the task to be queried. Passing a NULL handle results in the priority of the calling task being returned.

@return The priority of xTask.

Example usage: @code{c} void vAFunction( void ) { TaskHandle_t xHandle;

// Create a task, storing the handle. xTaskCreate( vTaskCode, “NAME”, STACK_SIZE, NULL, tskIDLE_PRIORITY, &xHandle );

// …

// Use the handle to obtain the priority of the created task. // It was created with tskIDLE_PRIORITY, but may have changed // it itself. if( uxTaskPriorityGet( xHandle ) != tskIDLE_PRIORITY ) { // The task has changed it’s priority. }

// …

// Is our priority higher than the created task? if( uxTaskPriorityGet( xHandle ) < uxTaskPriorityGet( NULL ) ) { // Our priority (obtained using NULL handle) is higher. } } @endcode \ingroup TaskCtrl