esp_idf_hal::sys
Function vRingbufferGetInfo

pub unsafe extern "C" fn vRingbufferGetInfo(
    xRingbuffer: *mut c_void,
    uxFree: *mut u32,
    uxRead: *mut u32,
    uxWrite: *mut u32,
    uxAcquire: *mut u32,
    uxItemsWaiting: *mut u32,
)

@brief Get information about ring buffer status

Get information of a ring buffer’s current status such as free/read/write/acquire pointer positions, and number of items waiting to be retrieved. Arguments can be set to NULL if they are not required.

@param[in] xRingbuffer Ring buffer handle @param[out] uxFree Pointer use to store free pointer position @param[out] uxRead Pointer use to store read pointer position @param[out] uxWrite Pointer use to store write pointer position @param[out] uxAcquire Pointer use to store acquire pointer position @param[out] uxItemsWaiting Pointer use to store number of items (bytes for byte buffer) waiting to be retrieved