esp_idf_hal::sys
Function vTaskPrioritySet

pub unsafe extern "C" fn vTaskPrioritySet(
    xTask: *mut tskTaskControlBlock,
    uxNewPriority: u32,
)

INCLUDE_vTaskPrioritySet must be defined as 1 for this function to be available. See the configuration section for more information.

Set the priority of any task.

A context switch will occur before the function returns if the priority being set is higher than the currently executing task.

@param xTask Handle to the task for which the priority is being set. Passing a NULL handle results in the priority of the calling task being set.

@param uxNewPriority The priority to which the task will be set.

Example usage: @code{c} void vAFunction( void ) { TaskHandle_t xHandle;

// Create a task, storing the handle. xTaskCreate( vTaskCode, “NAME”, STACK_SIZE, NULL, tskIDLE_PRIORITY, &xHandle );

// …

// Use the handle to raise the priority of the created task. vTaskPrioritySet( xHandle, tskIDLE_PRIORITY + 1 );

// …

// Use a NULL handle to raise our priority to the same value. vTaskPrioritySet( NULL, tskIDLE_PRIORITY + 1 ); } @endcode \ingroup TaskCtrl