esp_idf_hal::sys
Function wifi_prov_mgr_endpoint_create

pub unsafe extern "C" fn wifi_prov_mgr_endpoint_create(
    ep_name: *const u8,
) -> i32

@brief Create an additional endpoint and allocate internal resources for it

This API is to be called by the application if it wants to create an additional endpoint. All additional endpoints will be assigned UUIDs starting from 0xFF54 and so on in the order of execution.

protocomm handler for the created endpoint is to be registered later using wifi_prov_mgr_endpoint_register() after provisioning has started.

@note This API can only be called BEFORE provisioning is started

@note Additional endpoints can be used for configuring client provided parameters other than Wi-Fi credentials, that are necessary for the main application and hence must be set prior to starting the application

@note After session establishment, the additional endpoints must be targeted first by the client side application before sending Wi-Fi configuration, because once Wi-Fi configuration finishes the provisioning service is stopped and hence all endpoints are unregistered

@param[in] ep_name unique name of the endpoint

@return

ESP_OK : Success

ESP_FAIL : Failure