esp_idf_hal::sys
Function xQueuePeek

pub unsafe extern "C" fn xQueuePeek(
    xQueue: *mut QueueDefinition,
    pvBuffer: *mut c_void,
    xTicksToWait: u32,
) -> i32

Receive an item from a queue without removing the item from the queue. The item is received by copy so a buffer of adequate size must be provided. The number of bytes copied into the buffer was defined when the queue was created.

Successfully received items remain on the queue so will be returned again by the next call, or a call to xQueueReceive().

This macro must not be used in an interrupt service routine. See xQueuePeekFromISR() for an alternative that can be called from an interrupt service routine.

@param xQueue The handle to the queue from which the item is to be received.

@param pvBuffer Pointer to the buffer into which the received item will be copied.

@param xTicksToWait The maximum amount of time the task should block waiting for an item to receive should the queue be empty at the time of the call. The time is defined in tick periods so the constant portTICK_PERIOD_MS should be used to convert to real time if this is required. xQueuePeek() will return immediately if xTicksToWait is 0 and the queue is empty.

@return pdTRUE if an item was successfully received from the queue, otherwise pdFALSE.

Example usage: @code{c} struct AMessage { char ucMessageID; char ucData[ 20 ]; } xMessage;

QueueHandle_t xQueue;

// Task to create a queue and post a value. void vATask( void *pvParameters ) { struct AMessage *pxMessage;

// Create a queue capable of containing 10 pointers to AMessage structures. // These should be passed by pointer as they contain a lot of data. xQueue = xQueueCreate( 10, sizeof( struct AMessage * ) ); if( xQueue == 0 ) { // Failed to create the queue. }

// …

// Send a pointer to a struct AMessage object. Don’t block if the // queue is already full. pxMessage = & xMessage; xQueueSend( xQueue, ( void * ) &pxMessage, ( TickType_t ) 0 );

// … Rest of task code. }

// Task to peek the data from the queue. void vADifferentTask( void *pvParameters ) { struct AMessage *pxRxedMessage;

if( xQueue != 0 ) { // Peek a message on the created queue. Block for 10 ticks if a // message is not immediately available. if( xQueuePeek( xQueue, &( pxRxedMessage ), ( TickType_t ) 10 ) ) { // pcRxedMessage now points to the struct AMessage variable posted // by vATask, but the item still remains on the queue. } }

// … Rest of task code. } @endcode \ingroup QueueManagement