Function esp_idf_hal::sys::xStreamBufferSendCompletedFromISR

pub unsafe extern "C" fn xStreamBufferSendCompletedFromISR(
    xStreamBuffer: *mut StreamBufferDef_t,
    pxHigherPriorityTaskWoken: *mut i32,
) -> i32

@cond !DOC_EXCLUDE_HEADER_SECTION stream_buffer.h

@code{c} BaseType_t xStreamBufferSendCompletedFromISR( StreamBufferHandle_t xStreamBuffer, BaseType_t *pxHigherPriorityTaskWoken ); @endcode @endcond

For advanced users only.

The sbSEND_COMPLETED() macro is called from within the FreeRTOS APIs when data is sent to a message buffer or stream buffer. If there was a task that was blocked on the message or stream buffer waiting for data to arrive then the sbSEND_COMPLETED() macro sends a notification to the task to remove it from the Blocked state. xStreamBufferSendCompletedFromISR() does the same thing. It is provided to enable application writers to implement their own version of sbSEND_COMPLETED(), and MUST NOT BE USED AT ANY OTHER TIME.

See the example implemented in FreeRTOS/Demo/Minimal/MessageBufferAMP.c for additional information.

@param xStreamBuffer The handle of the stream buffer to which data was written.

@param pxHigherPriorityTaskWoken *pxHigherPriorityTaskWoken should be initialised to pdFALSE before it is passed into xStreamBufferSendCompletedFromISR(). If calling xStreamBufferSendCompletedFromISR() removes a task from the Blocked state, and the task has a priority above the priority of the currently running task, then *pxHigherPriorityTaskWoken will get set to pdTRUE indicating that a context switch should be performed before exiting the ISR.

@return If a task was removed from the Blocked state then pdTRUE is returned. Otherwise pdFALSE is returned.

@cond !DOC_SINGLE_GROUP \defgroup xStreamBufferSendCompletedFromISR xStreamBufferSendCompletedFromISR @endcond \ingroup StreamBufferManagement