Function esp_idf_hal::sys::xTaskResumeFromISR

pub unsafe extern "C" fn xTaskResumeFromISR(
    xTaskToResume: *mut tskTaskControlBlock,
) -> i32

@cond !DOC_EXCLUDE_HEADER_SECTION task. h @code{c} void xTaskResumeFromISR( TaskHandle_t xTaskToResume ); @endcode @endcond

INCLUDE_xTaskResumeFromISR must be defined as 1 for this function to be available. See the configuration section for more information.

An implementation of vTaskResume() that can be called from within an ISR.

A task that has been suspended by one or more calls to vTaskSuspend () will be made available for running again by a single call to xTaskResumeFromISR ().

xTaskResumeFromISR() should not be used to synchronise a task with an interrupt if there is a chance that the interrupt could arrive prior to the task being suspended - as this can lead to interrupts being missed. Use of a semaphore as a synchronisation mechanism would avoid this eventuality.

@param xTaskToResume Handle to the task being readied.

@return pdTRUE if resuming the task should result in a context switch, otherwise pdFALSE. This is used by the ISR to determine if a context switch may be required following the ISR.

@cond !DOC_SINGLE_GROUP \defgroup vTaskResumeFromISR vTaskResumeFromISR @endcond \ingroup TaskCtrl