esp_idf_hal::sys

Function xTimerPendFunctionCall

pub unsafe extern "C" fn xTimerPendFunctionCall(
    xFunctionToPend: Option<unsafe extern "C" fn(*mut c_void, u32)>,
    pvParameter1: *mut c_void,
    ulParameter2: u32,
    xTicksToWait: u32,
) -> i32

Used to defer the execution of a function to the RTOS daemon task (the timer service task, hence this function is implemented in timers.c and is prefixed with ‘Timer’).

@param xFunctionToPend The function to execute from the timer service/ daemon task. The function must conform to the PendedFunction_t prototype.

@param pvParameter1 The value of the callback function’s first parameter. The parameter has a void * type to allow it to be used to pass any type. For example, unsigned longs can be cast to a void *, or the void * can be used to point to a structure.

@param ulParameter2 The value of the callback function’s second parameter.

@param xTicksToWait Calling this function will result in a message being sent to the timer daemon task on a queue. xTicksToWait is the amount of time the calling task should remain in the Blocked state (so not using any processing time) for space to become available on the timer queue if the queue is found to be full.

@return pdPASS is returned if the message was successfully sent to the timer daemon task, otherwise pdFALSE is returned.