esp_idf_hal::sys

Struct esp_bt_controller_config_t

#[repr(C)]
pub struct esp_bt_controller_config_t {

    pub magic: u32,
    pub version: u32,
    pub controller_task_stack_size: u16,
    pub controller_task_prio: u8,
    pub controller_task_run_cpu: u8,
    pub bluetooth_mode: u8,
    pub ble_max_act: u8,
    pub sleep_mode: u8,
    pub sleep_clock: u8,
    pub ble_st_acl_tx_buf_nb: u8,
    pub ble_hw_cca_check: u8,
    pub ble_adv_dup_filt_max: u16,
    pub coex_param_en: bool,
    pub ce_len_type: u8,
    pub coex_use_hooks: bool,
    pub hci_tl_type: u8,
    pub hci_tl_funcs: *mut esp_bt_hci_tl_t,
    pub txant_dft: u8,
    pub rxant_dft: u8,
    pub txpwr_dft: u8,
    pub cfg_mask: u32,
    pub scan_duplicate_mode: u8,
    pub scan_duplicate_type: u8,
    pub normal_adv_size: u16,
    pub mesh_adv_size: u16,
    pub coex_phy_coded_tx_rx_time_limit: u8,
    pub hw_target_code: u32,
    pub slave_ce_len_min: u8,
    pub hw_recorrect_en: u8,
    pub cca_thresh: u8,
    pub scan_backoff_upperlimitmax: u16,
    pub dup_list_refresh_period: u16,
    pub ble_50_feat_supp: bool,
    pub ble_cca_mode: u8,
    pub ble_data_lenth_zero_aux: u8,
    pub ble_chan_ass_en: u8,
    pub ble_ping_en: u8,
    pub ble_llcp_disc_flag: u8,
    pub run_in_flash: bool,
    pub dtm_en: bool,
    pub enc_en: bool,
    pub qa_test: bool,
    pub connect_en: bool,
    pub scan_en: bool,
    pub ble_aa_check: bool,
    pub adv_en: bool,

}

@brief Bluetooth Controller config options @note 1. For parameters configurable through menuconfig, it is recommended to adjust them via the menuconfig interface. Please refer to menuconfig for details on the range and default values. 2. It is not recommended to modify the values for parameters which are not configurable through menuconfig.

Fields
magic: u32

< Magic number

version: u32

< version number of the defined structure

controller_task_stack_size: u16

< Bluetooth Controller task stack size in bytes

controller_task_prio: u8

< Bluetooth Controller task priority

controller_task_run_cpu: u8

< CPU number that Bluetooth Controller task runs on. Configurable in menuconfig.

0 - CPU 0 (default)

1 - CPU 1

bluetooth_mode: u8

< BLE mode

ble_max_act: u8

< The maximum number of BLE instance. Configurable in menuconfig.

Range: 1 - 10

Default: 6

sleep_mode: u8

< Modem sleep mode. Configurable in menuconfig.

0 - Disable (default)

1 - Enable

sleep_clock: u8

< Modem sleep clock source. Configurable in menuconfig.

ble_st_acl_tx_buf_nb: u8

< Static ACL TX buffer numbers. Configurable in menuconfig.

Range: 0 - 12

Default: 0

ble_hw_cca_check: u8

< Deprecated

ble_adv_dup_filt_max: u16

< The maximum number of extended duplicate scan filter. Configurable in menuconfig.

Range: 1 - 500

Default: 30

coex_param_en: bool

< Deprecated

ce_len_type: u8

< Connection event length determination method. Configurable in menuconfig.

0 - Original (default)

1 - use CE_LEN parameter from HCI commands

2 - Espressif vendor defined method

coex_use_hooks: bool

< Deprecated

hci_tl_type: u8

< HCI transport layer type. Configurable in menuconfig.

0 - URAT

1 - Virtual HCI (default)

hci_tl_funcs: *mut esp_bt_hci_tl_t

< HCI transport functions used. It must be set when hci_tl_type is UART.

txant_dft: u8

< Default TX antenna. Configurable in menuconfig.

0 - Antenna 0 (default)

1 - Antenna 1

rxant_dft: u8

< Default RX antenna. Configurable in menuconfig.

0 - Antenna 0 (default)

1 - Antenna 1

txpwr_dft: u8

< Default TX power. Please refer to esp_power_level_t for supported power level. Configurable in menuconfig.

Default : ESP_PWR_LVL_P9 +9 dBm.

cfg_mask: u32

< Configuration mask to set specific options

scan_duplicate_mode: u8

< Scan duplicate filtering mode. Configurable in menuconfig.

0 - Normal scan duplicate filtering mode (default)

1 - Special scan duplicate filtering mode for BLE Mesh

scan_duplicate_type: u8

< Scan duplicate filtering type. If scan_duplicate_mode is set to 1, this parameter will be ignored. Configurable in menuconfig.

0 - Filter scan duplicates by device address only (default)

1 - Filter scan duplicates by advertising data only, even if they originate from different devices.

2 - Filter scan duplicated by device address and advertising data.

normal_adv_size: u16

< Maximum number of devices in scan duplicate filtering list. Configurable in menuconfig.

Range: 10 - 1000

Default: 100

mesh_adv_size: u16

< Maximum number of Mesh advertising packets in scan duplicate filtering list. Configurable in menuconfig.

Range: 10 - 1000

Default: 100

coex_phy_coded_tx_rx_time_limit: u8

< Enable / disable the maximum TX/RX time limit for Coded-PHY connections in coexistence with Wi-Fi scenarios. Configurable in menuconfig.

0 - Disable (default)

1 - Enable

hw_target_code: u32

< Hardware target. Internal use only. Please do not modify this value.

slave_ce_len_min: u8

< Slave minimum connection event length: 5 slots. Please do not modify this value.

hw_recorrect_en: u8

< Enable / disable uncoded phy / coded phy hardware re-correction. Configurable in menuconfig.

cca_thresh: u8

< Absolute value of hardware-triggered CCA threshold. The CCA threshold is always negative. If the channel assessment result exceeds the CCA threshold (e.g. -75 dBm), indicating the channel is busy, the hardware will not transmit packets on that channel. Configurable in menuconfig.

Range: 20 dBm - 100 dBm

Default: 75 dBm

scan_backoff_upperlimitmax: u16

< Enable / disable active scan backoff. Configurable in menuconfig.

0 - Disable (default)

1 - Enable

dup_list_refresh_period: u16

< Scan duplicate filtering list refresh period in seconds. Configurable in menuconfig

Range: 0 - 100 seconds

Default: 0 second

ble_50_feat_supp: bool

< True if BLE 5.0 features are enabled; false otherwise. This option depends on whether the Host enable the 5.0 features.

ble_cca_mode: u8

< BLE CCA mode. Configurable in menuconfig

0 - Disable (default)

1 - Hardware-triggered CCA

2 - Software-based CCA (experimental)

ble_data_lenth_zero_aux: u8

< Enable / disable auxiliary packets when the extended ADV data length is zero. Configurable in menuconfig.

0 - Disable (default)

1 - Enable

ble_chan_ass_en: u8

< Enable / disable BLE channel assessment. Configurable in menuconfig.

0 - Disable

1 - Enable (default)

ble_ping_en: u8

< Enable / disable BLE ping procedure. Configurable in menuconfig.

0 - Disable

1 - Enable (default)

ble_llcp_disc_flag: u8

< Flag indicating whether the Controller disconnects after Instant Passed (0x28) error occurs. Configurable in menuconfig.

The Controller does not disconnect after Instant Passed (0x28) by default.

run_in_flash: bool

< True if the Controller code is in flash (flash model); false otherwise (default). Configurable in menuconfig.

dtm_en: bool

< True if the DTM feature is enabled; false otherwise (default). Configurable in menuconfig.

enc_en: bool

< True if the encryption feature is enabled (default); false otherwise. Configurable in menuconfig.

qa_test: bool

< True if the QA test feature is enabled; false otherwise (default). Configurable in menuconfig.

connect_en: bool

< True if the connection feature is enabled (default); false otherwise. Configurable in menuconfig.

scan_en: bool

< True if the scan feature is enabled (default); false otherwise. Configurable in menuconfig.

ble_aa_check: bool

< True if adds a verification step for the Access Address within the CONNECT_IND PDU; false otherwise. Configurable in menuconfig

adv_en: bool

< True if the ADV feature is enabled (default); false otherwise. Configurable in menuconfig.