esp_idf_hal::sys

Struct mbedtls_cipher_context_t

#[repr(C)]
pub struct mbedtls_cipher_context_t {
    pub private_cipher_info: *const mbedtls_cipher_info_t,
    pub private_key_bitlen: i32,
    pub private_operation: i32,
    pub private_add_padding: Option<unsafe extern "C" fn(*mut u8, usize, usize)>,
    pub private_get_padding: Option<unsafe extern "C" fn(*mut u8, usize, *mut usize, *mut usize) -> i32>,
    pub private_unprocessed_data: [u8; 16],
    pub private_unprocessed_len: usize,
    pub private_iv: [u8; 16],
    pub private_iv_size: usize,
    pub private_cipher_ctx: *mut c_void,
    pub private_cmac_ctx: *mut mbedtls_cmac_context_t,
}

Generic cipher context.

Fields
private_cipher_info: *const mbedtls_cipher_info_t

Information about the associated cipher.

private_key_bitlen: i32

Key length to use.

private_operation: i32

Operation that the key of the context has been initialized for.

private_add_padding: Option<unsafe extern "C" fn(*mut u8, usize, usize)>

Padding functions to use, if relevant for the specific cipher mode.

private_get_padding: Option<unsafe extern "C" fn(*mut u8, usize, *mut usize, *mut usize) -> i32>private_unprocessed_data: [u8; 16]

Buffer for input that has not been processed yet.

private_unprocessed_len: usize

Number of Bytes that have not been processed yet.

private_iv: [u8; 16]

Current IV or NONCE_COUNTER for CTR-mode, data unit (or sector) number for XTS-mode.

private_iv_size: usize

IV size in Bytes, for ciphers with variable-length IVs.

private_cipher_ctx: *mut c_void

The cipher-specific context.

private_cmac_ctx: *mut mbedtls_cmac_context_t

CMAC-specific context.