esp_idf_hal::sys
Struct mbedtls_x509_crt

#[repr(C)]
pub struct mbedtls_x509_crt {

    pub private_own_buffer: i32,
    pub raw: mbedtls_asn1_buf,
    pub tbs: mbedtls_asn1_buf,
    pub version: i32,
    pub serial: mbedtls_asn1_buf,
    pub sig_oid: mbedtls_asn1_buf,
    pub issuer_raw: mbedtls_asn1_buf,
    pub subject_raw: mbedtls_asn1_buf,
    pub issuer: mbedtls_asn1_named_data,
    pub subject: mbedtls_asn1_named_data,
    pub valid_from: mbedtls_x509_time,
    pub valid_to: mbedtls_x509_time,
    pub pk_raw: mbedtls_asn1_buf,
    pub pk: mbedtls_pk_context,
    pub issuer_id: mbedtls_asn1_buf,
    pub subject_id: mbedtls_asn1_buf,
    pub v3_ext: mbedtls_asn1_buf,
    pub subject_alt_names: mbedtls_asn1_sequence,
    pub subject_key_id: mbedtls_asn1_buf,
    pub authority_key_id: mbedtls_x509_authority,
    pub certificate_policies: mbedtls_asn1_sequence,
    pub private_ext_types: i32,
    pub private_ca_istrue: i32,
    pub private_max_pathlen: i32,
    pub private_key_usage: u32,
    pub ext_key_usage: mbedtls_asn1_sequence,
    pub private_ns_cert_type: u8,
    pub private_sig: mbedtls_asn1_buf,
    pub private_sig_md: u32,
    pub private_sig_pk: u32,
    pub private_sig_opts: *mut c_void,
    pub next: *mut mbedtls_x509_crt,

}

Container for an X.509 certificate. The certificate may be chained.

Some fields of this structure are publicly readable. Do not modify them except via Mbed TLS library functions: the effect of modifying those fields or the data that those fields points to is unspecified.

Fields
private_own_buffer: i32raw: mbedtls_asn1_buf

< The raw certificate data (DER).

tbs: mbedtls_asn1_buf

< The raw certificate body (DER). The part that is To Be Signed.

version: i32

< The X.509 version. (1=v1, 2=v2, 3=v3)

serial: mbedtls_asn1_buf

< Unique id for certificate issued by a specific CA.

sig_oid: mbedtls_asn1_buf

< Signature algorithm, e.g. sha1RSA

issuer_raw: mbedtls_asn1_buf

< The raw issuer data (DER). Used for quick comparison.

subject_raw: mbedtls_asn1_buf

< The raw subject data (DER). Used for quick comparison.

issuer: mbedtls_asn1_named_data

< The parsed issuer data (named information object).

subject: mbedtls_asn1_named_data

< The parsed subject data (named information object).

valid_from: mbedtls_x509_time

< Start time of certificate validity.

valid_to: mbedtls_x509_time

< End time of certificate validity.

pk_raw: mbedtls_asn1_bufpk: mbedtls_pk_context

< Container for the public key context.

issuer_id: mbedtls_asn1_buf

< Optional X.509 v2/v3 issuer unique identifier.

subject_id: mbedtls_asn1_buf

< Optional X.509 v2/v3 subject unique identifier.

v3_ext: mbedtls_asn1_buf

< Optional X.509 v3 extensions.

subject_alt_names: mbedtls_asn1_sequence

< Optional list of raw entries of Subject Alternative Names extension. These can be later parsed by mbedtls_x509_parse_subject_alt_name.

subject_key_id: mbedtls_asn1_buf

< Optional X.509 v3 extension subject key identifier.

authority_key_id: mbedtls_x509_authority

< Optional X.509 v3 extension authority key identifier.

certificate_policies: mbedtls_asn1_sequence

< Optional list of certificate policies (Only anyPolicy is printed and enforced, however the rest of the policies are still listed).

private_ext_types: i32private_ca_istrue: i32private_max_pathlen: i32private_key_usage: u32ext_key_usage: mbedtls_asn1_sequence

< Optional list of extended key usage OIDs.

private_ns_cert_type: u8private_sig: mbedtls_asn1_bufprivate_sig_md: u32private_sig_pk: u32private_sig_opts: *mut c_voidnext: *mut mbedtls_x509_crt

Next certificate in the linked list that constitutes the CA chain. \p NULL indicates the end of the list. Do not modify this field directly.