esp_idf_hal::sys

Struct sdmmc_host_t

#[repr(C)]
pub struct sdmmc_host_t {

    pub flags: u32,
    pub slot: i32,
    pub max_freq_khz: i32,
    pub io_voltage: f32,
    pub driver_strength: u32,
    pub current_limit: u32,
    pub init: Option<unsafe extern "C" fn() -> i32>,
    pub set_bus_width: Option<unsafe extern "C" fn(i32, usize) -> i32>,
    pub get_bus_width: Option<unsafe extern "C" fn(i32) -> usize>,
    pub set_bus_ddr_mode: Option<unsafe extern "C" fn(i32, bool) -> i32>,
    pub set_card_clk: Option<unsafe extern "C" fn(i32, u32) -> i32>,
    pub set_cclk_always_on: Option<unsafe extern "C" fn(i32, bool) -> i32>,
    pub do_transaction: Option<unsafe extern "C" fn(i32, *mut sdmmc_command_t) -> i32>,
    pub __bindgen_anon_1: sdmmc_host_t__bindgen_ty_1,
    pub io_int_enable: Option<unsafe extern "C" fn(i32) -> i32>,
    pub io_int_wait: Option<unsafe extern "C" fn(i32, u32) -> i32>,
    pub command_timeout_ms: i32,
    pub get_real_freq: Option<unsafe extern "C" fn(i32, *mut i32) -> i32>,
    pub input_delay_phase: u32,
    pub set_input_delay: Option<unsafe extern "C" fn(i32, u32) -> i32>,
    pub dma_aligned_buffer: *mut c_void,
    pub pwr_ctrl_handle: *mut sd_pwr_ctrl_drv_t,
    pub get_dma_info: Option<unsafe extern "C" fn(i32, *mut esp_dma_mem_info_t) -> i32>,
    pub check_buffer_alignment: Option<unsafe extern "C" fn(i32, *const c_void, usize) -> bool>,
    pub is_slot_set_to_uhs1: Option<unsafe extern "C" fn(i32, *mut bool) -> i32>,

}

SD/MMC Host description

This structure defines properties of SD/MMC host and functions of SD/MMC host which can be used by upper layers.

Fields
flags: u32

< flags defining host properties

slot: i32

< slot number, to be passed to host functions

max_freq_khz: i32

< max frequency supported by the host

io_voltage: f32

< I/O voltage used by the controller (voltage switching is not supported)

driver_strength: u32

< Driver Strength

current_limit: u32

< Current Limit

init: Option<unsafe extern "C" fn() -> i32>

< Host function to initialize the driver

set_bus_width: Option<unsafe extern "C" fn(i32, usize) -> i32>

< host function to set bus width

get_bus_width: Option<unsafe extern "C" fn(i32) -> usize>

< host function to get bus width

set_bus_ddr_mode: Option<unsafe extern "C" fn(i32, bool) -> i32>

< host function to set DDR mode

set_card_clk: Option<unsafe extern "C" fn(i32, u32) -> i32>

< host function to set card clock frequency

set_cclk_always_on: Option<unsafe extern "C" fn(i32, bool) -> i32>

< host function to set whether the clock is always enabled

do_transaction: Option<unsafe extern "C" fn(i32, *mut sdmmc_command_t) -> i32>

< host function to do a transaction

__bindgen_anon_1: sdmmc_host_t__bindgen_ty_1io_int_enable: Option<unsafe extern "C" fn(i32) -> i32>

< Host function to enable SDIO interrupt line

io_int_wait: Option<unsafe extern "C" fn(i32, u32) -> i32>

< Host function to wait for SDIO interrupt line to be active

command_timeout_ms: i32

< timeout, in milliseconds, of a single command. Set to 0 to use the default value.

get_real_freq: Option<unsafe extern "C" fn(i32, *mut i32) -> i32>

< Host function to provide real working freq, based on SDMMC controller setup

input_delay_phase: u32

< input delay phase, this will only take into effect when the host works in SDMMC_FREQ_HIGHSPEED or SDMMC_FREQ_52M. Driver will print out how long the delay is

set_input_delay: Option<unsafe extern "C" fn(i32, u32) -> i32>

< set input delay phase

dma_aligned_buffer: *mut c_void

< Leave it NULL. Reserved for cache aligned buffers for SDIO mode

pwr_ctrl_handle: *mut sd_pwr_ctrl_drv_t

< Power control handle

get_dma_info: Option<unsafe extern "C" fn(i32, *mut esp_dma_mem_info_t) -> i32>

< host function to dma memory information

check_buffer_alignment: Option<unsafe extern "C" fn(i32, *const c_void, usize) -> bool>

< Check if buffer meets alignment requirements

is_slot_set_to_uhs1: Option<unsafe extern "C" fn(i32, *mut bool) -> i32>

< host slot is set to uhs1 or not