esp_idf_hal::sys
Struct spi_bus_config_t

#[repr(C)]
pub struct spi_bus_config_t {

    pub __bindgen_anon_1: spi_bus_config_t__bindgen_ty_1,
    pub __bindgen_anon_2: spi_bus_config_t__bindgen_ty_2,
    pub sclk_io_num: i32,
    pub __bindgen_anon_3: spi_bus_config_t__bindgen_ty_3,
    pub __bindgen_anon_4: spi_bus_config_t__bindgen_ty_4,
    pub data4_io_num: i32,
    pub data5_io_num: i32,
    pub data6_io_num: i32,
    pub data7_io_num: i32,
    pub max_transfer_sz: i32,
    pub flags: u32,
    pub isr_cpu_id: u32,
    pub intr_flags: i32,

}

@brief This is a configuration structure for a SPI bus.

You can use this structure to specify the GPIO pins of the bus. Normally, the driver will use the GPIO matrix to route the signals. An exception is made when all signals either can be routed through the IO_MUX or are -1. In that case, the IO_MUX is used. On ESP32, using GPIO matrix will bring about 25ns of input delay, which may cause incorrect read for >40MHz speeds.

@note Be advised that the slave driver does not use the quadwp/quadhd lines and fields in spi_bus_config_t referring to these lines will be ignored and can thus safely be left uninitialized.

Fields
__bindgen_anon_1: spi_bus_config_t__bindgen_ty_1__bindgen_anon_2: spi_bus_config_t__bindgen_ty_2sclk_io_num: i32

< GPIO pin for SPI Clock signal, or -1 if not used.

__bindgen_anon_3: spi_bus_config_t__bindgen_ty_3__bindgen_anon_4: spi_bus_config_t__bindgen_ty_4data4_io_num: i32

< GPIO pin for spi data4 signal in octal mode, or -1 if not used.

data5_io_num: i32

< GPIO pin for spi data5 signal in octal mode, or -1 if not used.

data6_io_num: i32

< GPIO pin for spi data6 signal in octal mode, or -1 if not used.

data7_io_num: i32

< GPIO pin for spi data7 signal in octal mode, or -1 if not used.

max_transfer_sz: i32

< Maximum transfer size, in bytes. Defaults to 4092 if 0 when DMA enabled, or to SOC_SPI_MAXIMUM_BUFFER_SIZE if DMA is disabled.

flags: u32

< Abilities of bus to be checked by the driver. Or-ed value of SPICOMMON_BUSFLAG_* flags.

isr_cpu_id: u32

< Select cpu core to register SPI ISR.

intr_flags: i32

< Interrupt flag for the bus to set the priority, and IRAM attribute, see esp_intr_alloc.h. Note that the EDGE, INTRDISABLED attribute are ignored by the driver. Note that if ESP_INTR_FLAG_IRAM is set, ALL the callbacks of the driver, and their callee functions, should be put in the IRAM.