esp_idf_hal::sys
Struct spi_device_interface_config_t

#[repr(C)]
pub struct spi_device_interface_config_t {

    pub command_bits: u8,
    pub address_bits: u8,
    pub dummy_bits: u8,
    pub mode: u8,
    pub clock_source: u32,
    pub duty_cycle_pos: u16,
    pub cs_ena_pretrans: u16,
    pub cs_ena_posttrans: u8,
    pub clock_speed_hz: i32,
    pub input_delay_ns: i32,
    pub spics_io_num: i32,
    pub flags: u32,
    pub queue_size: i32,
    pub pre_cb: Option<unsafe extern "C" fn(_: *mut spi_transaction_t)>,
    pub post_cb: Option<unsafe extern "C" fn(_: *mut spi_transaction_t)>,

}

@brief This is a configuration for a SPI slave device that is connected to one of the SPI buses.

Fields
command_bits: u8

< Default amount of bits in command phase (0-16), used when SPI_TRANS_VARIABLE_CMD is not used, otherwise ignored.

address_bits: u8

< Default amount of bits in address phase (0-64), used when SPI_TRANS_VARIABLE_ADDR is not used, otherwise ignored.

dummy_bits: u8

< Amount of dummy bits to insert between address and data phase

mode: u8

< SPI mode, representing a pair of (CPOL, CPHA) configuration:

0: (0, 0)

1: (0, 1)

2: (1, 0)

3: (1, 1)

clock_source: u32

< Select SPI clock source, SPI_CLK_SRC_DEFAULT by default.

duty_cycle_pos: u16

< Duty cycle of positive clock, in 1/256th increments (128 = 50%/50% duty). Setting this to 0 (=not setting it) is equivalent to setting this to 128.

cs_ena_pretrans: u16

< Amount of SPI bit-cycles the cs should be activated before the transmission (0-16). This only works on half-duplex transactions.

cs_ena_posttrans: u8

< Amount of SPI bit-cycles the cs should stay active after the transmission (0-16)

clock_speed_hz: i32

< SPI clock speed in Hz. Derived from clock_source.

input_delay_ns: i32

< Maximum data valid time of slave. The time required between SCLK and MISO valid, including the possible clock delay from slave to master. The driver uses this value to give an extra delay before the MISO is ready on the line. Leave at 0 unless you know you need a delay. For better timing performance at high frequency (over 8MHz), itโs suggest to have the right value.

spics_io_num: i32

< CS GPIO pin for this device, or -1 if not used

flags: u32

< Bitwise OR of SPI_DEVICE_* flags

queue_size: i32

< Transaction queue size. This sets how many transactions can be โin the airโ (queued using spi_device_queue_trans but not yet finished using spi_device_get_trans_result) at the same time

pre_cb: Option<unsafe extern "C" fn(_: *mut spi_transaction_t)>

< Callback to be called before a transmission is started.

This callback is called within interrupt context should be in IRAM for best performance, see โTransferring Speedโ section in the SPI Master documentation for full details. If not, the callback may crash during flash operation when the driver is initialized with ESP_INTR_FLAG_IRAM.

post_cb: Option<unsafe extern "C" fn(_: *mut spi_transaction_t)>

< Callback to be called after a transmission has completed.

This callback is called within interrupt context should be in IRAM for best performance, see โTransferring Speedโ section in the SPI Master documentation for full details. If not, the callback may crash during flash operation when the driver is initialized with ESP_INTR_FLAG_IRAM.