esp_idf_hal::sys
Struct wifi_ap_record_t

#[repr(C)]
pub struct wifi_ap_record_t {

    pub bssid: [u8; 6],
    pub ssid: [u8; 33],
    pub primary: u8,
    pub second: u32,
    pub rssi: i8,
    pub authmode: u32,
    pub pairwise_cipher: u32,
    pub group_cipher: u32,
    pub ant: u32,
    pub _bitfield_align_1: [u32; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4]>,
    pub country: wifi_country_t,
    pub he_ap: wifi_he_ap_info_t,
    pub bandwidth: u8,
    pub vht_ch_freq1: u8,
    pub vht_ch_freq2: u8,

}

@brief Description of a Wi-Fi AP

Fields
bssid: [u8; 6]

< MAC address of AP

ssid: [u8; 33]

< SSID of AP

primary: u8

< Channel of AP

second: u32

< Secondary channel of AP

rssi: i8

< Signal strength of AP. Note that in some rare cases where signal strength is very strong, RSSI values can be slightly positive

authmode: u32

< Auth mode of AP

pairwise_cipher: u32

< Pairwise cipher of AP

group_cipher: u32

< Group cipher of AP

ant: u32

< Antenna used to receive beacon from AP

_bitfield_align_1: [u32; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 4]>country: wifi_country_t

< Country information of AP

he_ap: wifi_he_ap_info_t

< HE AP info

bandwidth: u8

< For AP 20 MHz this value is set to 1. For AP 40 MHz this value is set to 2. For AP 80 MHz this value is set to 3. For AP 160MHz this value is set to 4. For AP 80+80MHz this value is set to 5

vht_ch_freq1: u8

< This fields are used only AP bandwidth is 80 and 160 MHz, to transmit the center channel frequency of the BSS. For AP bandwidth is 80 + 80 MHz, it is the center channel frequency of the lower frequency segment.

vht_ch_freq2: u8

< This fields are used only AP bandwidth is 80 + 80 MHz, and is used to transmit the center channel frequency of the second segment.

Implementations

impl wifi_ap_record_t

pub fn phy_11b(&self) -> u32

pub fn set_phy_11b(&mut self, val: u32)

pub unsafe fn phy_11b_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11b_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_11g(&self) -> u32

pub fn set_phy_11g(&mut self, val: u32)

pub unsafe fn phy_11g_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11g_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_11n(&self) -> u32

pub fn set_phy_11n(&mut self, val: u32)

pub unsafe fn phy_11n_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11n_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_lr(&self) -> u32

pub fn set_phy_lr(&mut self, val: u32)

pub unsafe fn phy_lr_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_lr_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_11a(&self) -> u32

pub fn set_phy_11a(&mut self, val: u32)

pub unsafe fn phy_11a_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11a_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_11ac(&self) -> u32

pub fn set_phy_11ac(&mut self, val: u32)

pub unsafe fn phy_11ac_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11ac_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn phy_11ax(&self) -> u32

pub fn set_phy_11ax(&mut self, val: u32)

pub unsafe fn phy_11ax_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_phy_11ax_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn wps(&self) -> u32

pub fn set_wps(&mut self, val: u32)

pub unsafe fn wps_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_wps_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn ftm_responder(&self) -> u32

pub fn set_ftm_responder(&mut self, val: u32)

pub unsafe fn ftm_responder_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_ftm_responder_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn ftm_initiator(&self) -> u32

pub fn set_ftm_initiator(&mut self, val: u32)

pub unsafe fn ftm_initiator_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_ftm_initiator_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn reserved(&self) -> u32

pub fn set_reserved(&mut self, val: u32)

pub unsafe fn reserved_raw(this: *const wifi_ap_record_t) -> u32

pub unsafe fn set_reserved_raw(this: *mut wifi_ap_record_t, val: u32)

pub fn new_bitfield_1( phy_11b: u32, phy_11g: u32, phy_11n: u32, phy_lr: u32, phy_11a: u32, phy_11ac: u32, phy_11ax: u32, wps: u32, ftm_responder: u32, ftm_initiator: u32, reserved: u32, ) -> __BindgenBitfieldUnit<[u8; 4]>