esp_idf_hal::sys
Struct wifi_ftm_report_entry_t

#[repr(C)]
pub struct wifi_ftm_report_entry_t {
    pub dlog_token: u8,
    pub rssi: i8,
    pub rtt: u32,
    pub t1: u64,
    pub t2: u64,
    pub t3: u64,
    pub t4: u64,
}

@brief Structure representing a report entry for Fine Timing Measurement (FTM) in Wi-Fi.

This structure holds the information related to the FTM process between a Wi-Fi FTM Initiator and a Wi-Fi FTM Responder. FTM is used for precise distance measurement by timing the exchange of frames between devices.

Fields
dlog_token: u8

< Dialog Token of the FTM frame

rssi: i8

< RSSI of the FTM frame received

rtt: u32

< Round Trip Time in pSec with a peer

t1: u64

< Time of departure of FTM frame from FTM Responder in pSec

t2: u64

< Time of arrival of FTM frame at FTM Initiator in pSec

t3: u64

< Time of departure of ACK from FTM Initiator in pSec

t4: u64

< Time of arrival of ACK at FTM Responder in pSec