esp_idf_hal::sys
Struct wifi_pkt_rx_ctrl_t

#[repr(C)]
pub struct wifi_pkt_rx_ctrl_t {
    pub _bitfield_align_1: [u32; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 48]>,
}

@brief Received packet radio metadata header, this is the common header at the beginning of all promiscuous mode RX callback buffers

Fields
_bitfield_align_1: [u32; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 48]>
Implementations

impl wifi_pkt_rx_ctrl_t

pub fn rssi(&self) -> i32

pub fn set_rssi(&mut self, val: i32)

pub unsafe fn rssi_raw(this: *const wifi_pkt_rx_ctrl_t) -> i32

pub unsafe fn set_rssi_raw(this: *mut wifi_pkt_rx_ctrl_t, val: i32)

pub fn rate(&self) -> u32

pub fn set_rate(&mut self, val: u32)

pub unsafe fn rate_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_rate_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn sig_mode(&self) -> u32

pub fn set_sig_mode(&mut self, val: u32)

pub unsafe fn sig_mode_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_sig_mode_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn mcs(&self) -> u32

pub fn set_mcs(&mut self, val: u32)

pub unsafe fn mcs_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_mcs_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn cwb(&self) -> u32

pub fn set_cwb(&mut self, val: u32)

pub unsafe fn cwb_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_cwb_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn smoothing(&self) -> u32

pub fn set_smoothing(&mut self, val: u32)

pub unsafe fn smoothing_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_smoothing_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn not_sounding(&self) -> u32

pub fn set_not_sounding(&mut self, val: u32)

pub unsafe fn not_sounding_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_not_sounding_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn aggregation(&self) -> u32

pub fn set_aggregation(&mut self, val: u32)

pub unsafe fn aggregation_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_aggregation_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn stbc(&self) -> u32

pub fn set_stbc(&mut self, val: u32)

pub unsafe fn stbc_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_stbc_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn fec_coding(&self) -> u32

pub fn set_fec_coding(&mut self, val: u32)

pub unsafe fn fec_coding_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_fec_coding_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn sgi(&self) -> u32

pub fn set_sgi(&mut self, val: u32)

pub unsafe fn sgi_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_sgi_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn ampdu_cnt(&self) -> u32

pub fn set_ampdu_cnt(&mut self, val: u32)

pub unsafe fn ampdu_cnt_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_ampdu_cnt_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn channel(&self) -> u32

pub fn set_channel(&mut self, val: u32)

pub unsafe fn channel_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_channel_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn secondary_channel(&self) -> u32

pub fn set_secondary_channel(&mut self, val: u32)

pub unsafe fn secondary_channel_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_secondary_channel_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn timestamp(&self) -> u32

pub fn set_timestamp(&mut self, val: u32)

pub unsafe fn timestamp_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_timestamp_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn noise_floor(&self) -> i32

pub fn set_noise_floor(&mut self, val: i32)

pub unsafe fn noise_floor_raw(this: *const wifi_pkt_rx_ctrl_t) -> i32

pub unsafe fn set_noise_floor_raw(this: *mut wifi_pkt_rx_ctrl_t, val: i32)

pub fn ant(&self) -> u32

pub fn set_ant(&mut self, val: u32)

pub unsafe fn ant_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_ant_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn sig_len(&self) -> u32

pub fn set_sig_len(&mut self, val: u32)

pub unsafe fn sig_len_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_sig_len_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)

pub fn rx_state(&self) -> u32

pub fn set_rx_state(&mut self, val: u32)

pub unsafe fn rx_state_raw(this: *const wifi_pkt_rx_ctrl_t) -> u32

pub unsafe fn set_rx_state_raw(this: *mut wifi_pkt_rx_ctrl_t, val: u32)