Struct esp_idf_hal::sys::wifi_sta_config_t

#[repr(C)]
pub struct wifi_sta_config_t {

    pub ssid: [u8; 32],
    pub password: [u8; 64],
    pub scan_method: u32,
    pub bssid_set: bool,
    pub bssid: [u8; 6],
    pub channel: u8,
    pub listen_interval: u16,
    pub sort_method: u32,
    pub threshold: wifi_scan_threshold_t,
    pub pmf_cfg: wifi_pmf_config_t,
    pub _bitfield_align_1: [u32; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4]>,
    pub sae_pwe_h2e: u32,
    pub sae_pk_mode: u32,
    pub failure_retry_cnt: u8,
    pub _bitfield_align_2: [u32; 0],
    pub _bitfield_2: __BindgenBitfieldUnit<[u8; 4]>,
    pub sae_h2e_identifier: [u8; 32],

}

@brief STA configuration settings for the device

Fields
ssid: [u8; 32]

< SSID of target AP.

password: [u8; 64]

< Password of target AP.

scan_method: u32

< do all channel scan or fast scan

bssid_set: bool

< whether set MAC address of target AP or not. Generally, station_config.bssid_set needs to be 0; and it needs to be 1 only when users need to check the MAC address of the AP.

bssid: [u8; 6]

< MAC address of target AP

channel: u8

< channel of target AP. Set to 1~13 to scan starting from the specified channel before connecting to AP. If the channel of AP is unknown, set it to 0.

listen_interval: u16

< Listen interval for ESP32 station to receive beacon when WIFI_PS_MAX_MODEM is set. Units: AP beacon intervals. Defaults to 3 if set to 0.

sort_method: u32

< sort the connect AP in the list by rssi or security mode

threshold: wifi_scan_threshold_t

< When scan_threshold is set, only APs which have an auth mode that is more secure than the selected auth mode and a signal stronger than the minimum RSSI will be used.

pmf_cfg: wifi_pmf_config_t

< Configuration for Protected Management Frame. Will be advertised in RSN Capabilities in RSN IE.

_bitfield_align_1: [u32; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 4]>sae_pwe_h2e: u32

< Configuration for SAE PWE derivation method

sae_pk_mode: u32

< Configuration for SAE-PK (Public Key) Authentication method

failure_retry_cnt: u8

< Number of connection retries station will do before moving to next AP. scan_method should be set as WIFI_ALL_CHANNEL_SCAN to use this config. Note: Enabling this may cause connection time to increase incase best AP doesn’t behave properly.

_bitfield_align_2: [u32; 0]_bitfield_2: __BindgenBitfieldUnit<[u8; 4]>sae_h2e_identifier: [u8; 32]

< Password identifier for H2E. this needs to be null terminated string

Implementations

impl wifi_sta_config_t

pub fn rm_enabled(&self) -> u32

pub fn set_rm_enabled(&mut self, val: u32)

pub fn btm_enabled(&self) -> u32

pub fn set_btm_enabled(&mut self, val: u32)

pub fn mbo_enabled(&self) -> u32

pub fn set_mbo_enabled(&mut self, val: u32)

pub fn ft_enabled(&self) -> u32

pub fn set_ft_enabled(&mut self, val: u32)

pub fn owe_enabled(&self) -> u32

pub fn set_owe_enabled(&mut self, val: u32)

pub fn transition_disable(&self) -> u32

pub fn set_transition_disable(&mut self, val: u32)

pub fn reserved(&self) -> u32

pub fn set_reserved(&mut self, val: u32)

pub fn new_bitfield_1( rm_enabled: u32, btm_enabled: u32, mbo_enabled: u32, ft_enabled: u32, owe_enabled: u32, transition_disable: u32, reserved: u32, ) -> __BindgenBitfieldUnit<[u8; 4]>

pub fn he_dcm_set(&self) -> u32

pub fn set_he_dcm_set(&mut self, val: u32)

pub fn he_dcm_max_constellation_tx(&self) -> u32

pub fn set_he_dcm_max_constellation_tx(&mut self, val: u32)

pub fn he_dcm_max_constellation_rx(&self) -> u32

pub fn set_he_dcm_max_constellation_rx(&mut self, val: u32)

pub fn he_mcs9_enabled(&self) -> u32

pub fn set_he_mcs9_enabled(&mut self, val: u32)

pub fn he_su_beamformee_disabled(&self) -> u32

pub fn set_he_su_beamformee_disabled(&mut self, val: u32)

pub fn he_trig_su_bmforming_feedback_disabled(&self) -> u32

pub fn set_he_trig_su_bmforming_feedback_disabled(&mut self, val: u32)

pub fn he_trig_mu_bmforming_partial_feedback_disabled(&self) -> u32

pub fn set_he_trig_mu_bmforming_partial_feedback_disabled(&mut self, val: u32)

pub fn he_trig_cqi_feedback_disabled(&self) -> u32

pub fn set_he_trig_cqi_feedback_disabled(&mut self, val: u32)

pub fn he_reserved(&self) -> u32

pub fn set_he_reserved(&mut self, val: u32)

pub fn new_bitfield_2( he_dcm_set: u32, he_dcm_max_constellation_tx: u32, he_dcm_max_constellation_rx: u32, he_mcs9_enabled: u32, he_su_beamformee_disabled: u32, he_trig_su_bmforming_feedback_disabled: u32, he_trig_mu_bmforming_partial_feedback_disabled: u32, he_trig_cqi_feedback_disabled: u32, he_reserved: u32, ) -> __BindgenBitfieldUnit<[u8; 4]>