Type Alias esp_idf_hal::sys::esp_sntp_config_t

pub type esp_sntp_config_t = esp_sntp_config;

@brief SNTP configuration struct

Aliased Type

struct esp_sntp_config_t {
    pub smooth_sync: bool,
    pub server_from_dhcp: bool,
    pub wait_for_sync: bool,
    pub start: bool,
    pub sync_cb: Option<unsafe extern "C" fn(_: *mut timeval)>,
    pub renew_servers_after_new_IP: bool,
    pub ip_event_to_renew: u32,
    pub index_of_first_server: usize,
    pub num_of_servers: usize,
    pub servers: [*const i8; 1],
}

Fields
smooth_sync: bool

< set to true if smooth sync required

server_from_dhcp: bool

< set to true to request NTP server config from DHCP

wait_for_sync: bool

< if true, we create a semaphore to signal time sync event

start: bool

< set to true to automatically start the SNTP service

sync_cb: Option<unsafe extern "C" fn(_: *mut timeval)>

< optionally sets callback function on time sync event

renew_servers_after_new_IP: bool

< this is used to refresh server list if NTP provided by DHCP (which cleans other pre-configured servers)

ip_event_to_renew: u32

< set the IP event id on which we refresh server list (if renew_servers_after_new_IP=true)

index_of_first_server: usize

< refresh server list after this server (if renew_servers_after_new_IP=true)

num_of_servers: usize

< number of preconfigured NTP servers

servers: [*const i8; 1]

< list of servers