Type Alias esp_idf_hal::sys::i2s_config_t

pub type i2s_config_t = i2s_driver_config_t;

@brief I2S driver configuration parameters

Aliased Type

struct i2s_config_t {

    pub mode: u32,
    pub sample_rate: u32,
    pub bits_per_sample: u32,
    pub channel_format: u32,
    pub communication_format: u32,
    pub intr_alloc_flags: i32,
    pub dma_buf_count: i32,
    pub dma_buf_len: i32,
    pub use_apll: bool,
    pub tx_desc_auto_clear: bool,
    pub fixed_mclk: i32,
    pub mclk_multiple: u32,
    pub bits_per_chan: u32,
    pub chan_mask: i32,
    pub total_chan: u32,
    pub left_align: bool,
    pub big_edin: bool,
    pub bit_order_msb: bool,
    pub skip_msk: bool,

}

Fields
mode: u32

< I2S work mode

sample_rate: u32

< I2S sample rate

bits_per_sample: u32

< I2S sample bits in one channel

channel_format: u32

< I2S channel format.

communication_format: u32

< I2S communication format

intr_alloc_flags: i32

< Flags used to allocate the interrupt. One or multiple (ORred) ESP_INTR_FLAG_* values. See esp_intr_alloc.h for more info

dma_buf_count: i32

< The total number of DMA buffers to receive/transmit data. A descriptor includes some information such as buffer address, the address of the next descriptor, and the buffer length. Since one descriptor points to one buffer, therefore, ‘dma_desc_num’ can be interpreted as the total number of DMA buffers used to store data from DMA interrupt. Notice that these buffers are internal to’i2s_read’ and descriptors are created automatically inside of the I2S driver. Users only need to set the buffer number while the length is derived from the parameter described below.

dma_buf_len: i32

< Number of frames in a DMA buffer. A frame means the data of all channels in a WS cycle. The real_dma_buf_size = dma_buf_len * chan_num * bits_per_chan / 8. For example, if two channels in stereo mode (i.e., ‘channel_format’ is set to ‘I2S_CHANNEL_FMT_RIGHT_LEFT’) are active, and each channel transfers 32 bits (i.e., ‘bits_per_sample’ is set to ‘I2S_BITS_PER_CHAN_32BIT’), then the total number of bytes of a frame is ‘channel_format’ * ‘bits_per_sample’ = 2 * 32 / 8 = 8 bytes. We assume that the current ‘dma_buf_len’ is 100, then the real length of the DMA buffer is 8 * 100 = 800 bytes. Note that the length of an internal real DMA buffer shouldn’t be greater than 4092.

use_apll: bool

< I2S using APLL as main I2S clock, enable it to get accurate clock

tx_desc_auto_clear: bool

< I2S auto clear tx descriptor if there is underflow condition (helps in avoiding noise in case of data unavailability)

fixed_mclk: i32

< I2S using fixed MCLK output. If use_apll = true and fixed_mclk > 0, then the clock output for i2s is fixed and equal to the fixed_mclk value. If fixed_mclk set, mclk_multiple won’t take effect

mclk_multiple: u32

< The multiple of I2S master clock(MCLK) to sample rate

bits_per_chan: u32

< I2S total bits in one channel， only take effect when larger than ‘bits_per_sample’, default ‘0’ means equal to ‘bits_per_sample’

chan_mask: i32

< I2S active channel bit mask, set value in i2s_channel_t to enable specific channel, the bit map of active channel can not exceed (0x1<<total_chan).

total_chan: u32

< I2S Total number of channels. If it is smaller than the biggest active channel number, it will be set to this number automatically.

left_align: bool

< Set to enable left alignment

big_edin: bool

< Set to enable big edin

bit_order_msb: bool

< Set to enable msb order

skip_msk: bool

< Set to enable skip mask. If it is enabled, only the data of the enabled channels will be sent, otherwise all data stored in DMA TX buffer will be sent