esp_idf_hal::sys

Type Alias mbedtls_psa_stats_t

pub type mbedtls_psa_stats_t = mbedtls_psa_stats_s;

\brief Statistics about resource consumption related to the PSA keystore.

\note The content of this structure is not part of the stable API and ABI of Mbed TLS and may change arbitrarily from version to version.

Aliased Type

#[repr(C)]
pub struct mbedtls_psa_stats_t {
    pub private_volatile_slots: usize,
    pub private_persistent_slots: usize,
    pub private_external_slots: usize,
    pub private_half_filled_slots: usize,
    pub private_cache_slots: usize,
    pub private_empty_slots: usize,
    pub private_locked_slots: usize,
    pub private_max_open_internal_key_id: u32,
    pub private_max_open_external_key_id: u32,
}

Fields
private_volatile_slots: usize

Number of slots containing key material for a volatile key.

private_persistent_slots: usize

Number of slots containing key material for a key which is in internal persistent storage.

private_external_slots: usize

Number of slots containing a reference to a key in a secure element.

private_half_filled_slots: usize

Number of slots which are occupied, but do not contain key material yet.

private_cache_slots: usize

Number of slots that contain cache data.

private_empty_slots: usize

Number of slots that are not used for anything.

private_locked_slots: usize

Number of slots that are locked.

private_max_open_internal_key_id: u32

Largest key id value among open keys in internal persistent storage.

private_max_open_external_key_id: u32

Largest key id value among open keys in secure elements.