Type Alias esp_idf_hal::sys::mbedtls_x509_name

pub type mbedtls_x509_name = mbedtls_asn1_named_data;

Container for ASN1 named information objects. It allows for Relative Distinguished Names (e.g. cn=localhost,ou=code,etc.).

Aliased Type

struct mbedtls_x509_name {
    pub oid: mbedtls_asn1_buf,
    pub val: mbedtls_asn1_buf,
    pub next: *mut mbedtls_asn1_named_data,
    pub private_next_merged: u8,
}

Fields
oid: mbedtls_asn1_buf

< The object identifier.

val: mbedtls_asn1_buf

< The named value.

next: *mut mbedtls_asn1_named_data

The next entry in the sequence.

The details of memory management for named data sequences are not documented and may change in future versions. Set this field to \p NULL when initializing a structure, and do not modify it except via Mbed TLS library functions.

private_next_merged: u8