esp_idf_hal::sys

Type Alias mesh_event_child_disconnected_t

pub type mesh_event_child_disconnected_t = wifi_event_ap_stadisconnected_t;

@brief Child disconnected information

Aliased Type

#[repr(C)]
pub struct mesh_event_child_disconnected_t {
    pub mac: [u8; 6],
    pub aid: u8,
    pub is_mesh_child: bool,
    pub reason: u16,
}

Fields
mac: [u8; 6]

< MAC address of the station disconnects from the soft-AP

aid: u8

< AID that the Soft-AP assigned to the disconnected station

is_mesh_child: bool

< Flag indicating whether the disconnected station is a mesh child

reason: u16

< Disconnection reason