esp_idf_hal::sys
Type Alias psa_pake_operation_t

pub type psa_pake_operation_t = psa_pake_operation_s;

The type of the state data structure for PAKE operations.

Before calling any function on a PAKE operation object, the application must initialize it by any of the following means:

Set the structure to all-bits-zero, for example: \code psa_pake_operation_t operation; memset(&operation, 0, sizeof(operation)); \endcode

Initialize the structure to logical zero values, for example: \code psa_pake_operation_t operation = {0}; \endcode

Initialize the structure to the initializer #PSA_PAKE_OPERATION_INIT, for example: \code psa_pake_operation_t operation = PSA_PAKE_OPERATION_INIT; \endcode

Assign the result of the function psa_pake_operation_init() to the structure, for example: \code psa_pake_operation_t operation; operation = psa_pake_operation_init(); \endcode

This is an implementation-defined \c struct. Applications should not make any assumptions about the content of this structure. Implementation details can change in future versions without notice.

Aliased Type

struct psa_pake_operation_t {
    pub private_id: u32,
    pub private_alg: u32,
    pub private_primitive: u32,
    pub private_stage: u8,
    pub private_computation_stage: psa_pake_operation_s__bindgen_ty_1,
    pub private_data: psa_pake_operation_s__bindgen_ty_2,
}

Fields
private_id: u32private_alg: u32private_primitive: u32private_stage: u8private_computation_stage: psa_pake_operation_s__bindgen_ty_1private_data: psa_pake_operation_s__bindgen_ty_2