esp_idf_hal::sys
Type Alias psa_verify_hash_interruptible_operation_t

pub type psa_verify_hash_interruptible_operation_t = psa_verify_hash_interruptible_operation_s;

The type of the state data structure for interruptible hash verification operations.

Before calling any function on a sign hash operation object, the application must initialize it by any of the following means:

Set the structure to all-bits-zero, for example: \code psa_verify_hash_interruptible_operation_t operation; memset(&operation, 0, sizeof(operation)); \endcode

Initialize the structure to logical zero values, for example: \code psa_verify_hash_interruptible_operation_t operation = {0}; \endcode

Initialize the structure to the initializer #PSA_VERIFY_HASH_INTERRUPTIBLE_OPERATION_INIT, for example: \code psa_verify_hash_interruptible_operation_t operation = PSA_VERIFY_HASH_INTERRUPTIBLE_OPERATION_INIT; \endcode

Assign the result of the function psa_verify_hash_interruptible_operation_init() to the structure, for example: \code psa_verify_hash_interruptible_operation_t operation; operation = psa_verify_hash_interruptible_operation_init(); \endcode

This is an implementation-defined \c struct. Applications should not make any assumptions about the content of this structure. Implementation details can change in future versions without notice.

Aliased Type

struct psa_verify_hash_interruptible_operation_t {
    pub private_id: u32,
    pub private_ctx: psa_driver_verify_hash_interruptible_context_t,
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1]>,
    pub private_num_ops: u32,
}

Fields
private_id: u32private_ctx: psa_driver_verify_hash_interruptible_context_t_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 1]>private_num_ops: u32