Type Alias esp_idf_hal::sys::wifi_prov_config_handlers_t

pub type wifi_prov_config_handlers_t = wifi_prov_config_handlers;

@brief Internal handlers for receiving and responding to protocomm requests from master

This is to be passed as priv_data for protocomm request handler (refer to wifi_prov_config_data_handler()) when calling protocomm_add_endpoint().

Aliased Type

struct wifi_prov_config_handlers_t {
    pub get_status_handler: Option<unsafe extern "C" fn(_: *mut wifi_prov_config_get_data_t, _: *mut *mut wifi_prov_ctx) -> i32>,
    pub set_config_handler: Option<unsafe extern "C" fn(_: *const wifi_prov_config_set_data_t, _: *mut *mut wifi_prov_ctx) -> i32>,
    pub apply_config_handler: Option<unsafe extern "C" fn(_: *mut *mut wifi_prov_ctx) -> i32>,
    pub ctx: *mut wifi_prov_ctx,
}

Fields
get_status_handler: Option<unsafe extern "C" fn(_: *mut wifi_prov_config_get_data_t, _: *mut *mut wifi_prov_ctx) -> i32>

Handler function called when connection status of the slave (in WiFi station mode) is requested

set_config_handler: Option<unsafe extern "C" fn(_: *const wifi_prov_config_set_data_t, _: *mut *mut wifi_prov_ctx) -> i32>

Handler function called when WiFi connection configuration (eg. AP SSID, password, etc.) of the slave (in WiFi station mode) is to be set to user provided values

apply_config_handler: Option<unsafe extern "C" fn(_: *mut *mut wifi_prov_ctx) -> i32>

Handler function for applying the configuration that was set in set_config_handler. After applying the station may get connected to the AP or may fail to connect. The slave must be ready to convey the updated connection status information when get_status_handler is invoked again by the master.

ctx: *mut wifi_prov_ctx

Context pointer to be passed to above handler functions upon invocation