esp_idf_hal::sys

Union ble_gap_event__bindgen_ty_1

#[repr(C)]
pub union ble_gap_event__bindgen_ty_1 {

    pub connect: ble_gap_event__bindgen_ty_1__bindgen_ty_1,
    pub link_estab: ble_gap_event__bindgen_ty_1__bindgen_ty_2,
    pub disconnect: ble_gap_event__bindgen_ty_1__bindgen_ty_3,
    pub disc: ble_gap_disc_desc,
    pub disc_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_4,
    pub adv_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_5,
    pub conn_update: ble_gap_event__bindgen_ty_1__bindgen_ty_6,
    pub conn_update_req: ble_gap_event__bindgen_ty_1__bindgen_ty_7,
    pub term_failure: ble_gap_event__bindgen_ty_1__bindgen_ty_8,
    pub enc_change: ble_gap_event__bindgen_ty_1__bindgen_ty_9,
    pub passkey: ble_gap_event__bindgen_ty_1__bindgen_ty_10,
    pub notify_rx: ble_gap_event__bindgen_ty_1__bindgen_ty_11,
    pub notify_tx: ble_gap_event__bindgen_ty_1__bindgen_ty_12,
    pub subscribe: ble_gap_event__bindgen_ty_1__bindgen_ty_13,
    pub mtu: ble_gap_event__bindgen_ty_1__bindgen_ty_14,
    pub identity_resolved: ble_gap_event__bindgen_ty_1__bindgen_ty_15,
    pub repeat_pairing: ble_gap_repeat_pairing,
    pub phy_updated: ble_gap_event__bindgen_ty_1__bindgen_ty_16,
    pub pairing_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_17,
    pub authorize: ble_gap_event__bindgen_ty_1__bindgen_ty_18,
    pub reattempt_cnt: ble_gap_event__bindgen_ty_1__bindgen_ty_19,
    pub dtm_state: ble_gap_event__bindgen_ty_1__bindgen_ty_20,
    pub data_len_chg: ble_gap_event__bindgen_ty_1__bindgen_ty_21,
    pub rd_all_rem_feat: ble_gap_event__bindgen_ty_1__bindgen_ty_22,

}

A discriminated union containing additional details concerning the GAP event. The ‘type’ field indicates which member of the union is valid.

Fields
connect: ble_gap_event__bindgen_ty_1__bindgen_ty_1link_estab: ble_gap_event__bindgen_ty_1__bindgen_ty_2disconnect: ble_gap_event__bindgen_ty_1__bindgen_ty_3disc: ble_gap_disc_desc

Represents an advertising report received during a discovery procedure. Valid for the following event types: o BLE_GAP_EVENT_DISC

disc_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_4adv_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_5conn_update: ble_gap_event__bindgen_ty_1__bindgen_ty_6conn_update_req: ble_gap_event__bindgen_ty_1__bindgen_ty_7term_failure: ble_gap_event__bindgen_ty_1__bindgen_ty_8enc_change: ble_gap_event__bindgen_ty_1__bindgen_ty_9passkey: ble_gap_event__bindgen_ty_1__bindgen_ty_10notify_rx: ble_gap_event__bindgen_ty_1__bindgen_ty_11notify_tx: ble_gap_event__bindgen_ty_1__bindgen_ty_12subscribe: ble_gap_event__bindgen_ty_1__bindgen_ty_13mtu: ble_gap_event__bindgen_ty_1__bindgen_ty_14identity_resolved: ble_gap_event__bindgen_ty_1__bindgen_ty_15repeat_pairing: ble_gap_repeat_pairing

Represents a peer’s attempt to pair despite a bond already existing. The application has two options for handling this event type: o Retry: Return BLE_GAP_REPEAT_PAIRING_RETRY after deleting the conflicting bond. The stack will verify the bond has been deleted and continue the pairing procedure. If the bond is still present, this event will be reported again. o Ignore: Return BLE_GAP_REPEAT_PAIRING_IGNORE. The stack will silently ignore the pairing request.

Valid for the following event types: o BLE_GAP_EVENT_REPEAT_PAIRING

phy_updated: ble_gap_event__bindgen_ty_1__bindgen_ty_16pairing_complete: ble_gap_event__bindgen_ty_1__bindgen_ty_17authorize: ble_gap_event__bindgen_ty_1__bindgen_ty_18reattempt_cnt: ble_gap_event__bindgen_ty_1__bindgen_ty_19dtm_state: ble_gap_event__bindgen_ty_1__bindgen_ty_20data_len_chg: ble_gap_event__bindgen_ty_1__bindgen_ty_21rd_all_rem_feat: ble_gap_event__bindgen_ty_1__bindgen_ty_22