esp_idf_hal::task::embassy_sync
Struct EspRawMutex

pub struct EspRawMutex(/* private fields */);

A mutex that allows borrowing data across executors but NOT accross interrupts.

Safety

This mutex is safe to share between different executors.

Implementations

impl EspRawMutex

pub const fn new() -> Self

Create a new EspRawMutex.