esp_idf_hal::task::notification

Struct Notifier

pub struct Notifier(/* private fields */);

Implementations

impl Notifier

pub unsafe fn notify(&self, notification: NonZeroU32) -> (bool, bool)

Safety

Care should be taken to ensure that Notifier does not outlive the task in which the Notification that produced it was created.

If that happens, a dangling pointer instead of proper task handle will be passed to task::notify, which will result in memory corruption.

pub unsafe fn notify_and_yield(&self, notification: NonZeroU32) -> bool

Safety

Care should be taken to ensure that Notifier does not outlive the task in which the Notification that produced it was created.

If that happens, a dangling pointer instead of proper task handle will be passed to task::notify_and_yield, which will result in memory corruption.