esp_idf_hal::uart
Struct UartRxDriver

pub struct UartRxDriver<'d> { /* private fields */ }

Serial receiver

Implementations

impl<'d> UartRxDriver<'d>

pub fn new<UART: Uart>( uart: impl Peripheral<P = UART> + 'd, rx: impl Peripheral<P = impl InputPin> + 'd, cts: Option<impl Peripheral<P = impl InputPin> + 'd>, rts: Option<impl Peripheral<P = impl OutputPin> + 'd>, config: &Config, ) -> Result<Self, EspError>

Create a new serial receiver

pub fn event_queue(&self) -> Option<&Queue<UartEvent>>

Retrieves the event queue for this UART. Returns None if the config specified 0 for queue_size.

pub fn change_stop_bits(&self, stop_bits: StopBits) -> Result<&Self, EspError>

Change the number of stop bits

pub fn stop_bits(&self) -> Result<StopBits, EspError>

Returns the current number of stop bits

pub fn change_data_bits(&self, data_bits: DataBits) -> Result<&Self, EspError>

Change the number of data bits

pub fn data_bits(&self) -> Result<DataBits, EspError>

Return the current number of data bits

pub fn change_parity(&self, parity: Parity) -> Result<&Self, EspError>

Change the type of parity checking

pub fn parity(&self) -> Result<Parity, EspError>

Returns the current type of parity checking

pub fn change_baudrate<T: Into<Hertz> + Copy>( &self, baudrate: T, ) -> Result<&Self, EspError>

Change the baudrate.

Will automatically select the clock source. When possible the reference clock (1MHz) will be used, because this is constant when the clock source/frequency changes. However if one of the clock frequencies is below 10MHz or if the baudrate is above the reference clock or if the baudrate cannot be set within 1.5% then use the APB clock.

pub fn baudrate(&self) -> Result<Hertz, EspError>

Returns the current baudrate

pub fn read(&self, buf: &mut [u8], delay: TickType_t) -> Result<usize, EspError>

Read multiple bytes into a slice; block until specified timeout Returns:

Ok(0) if the buffer is of length 0

Ok(n) if n bytes were read, where n is > 0

Err(EspError::Timeout) if no bytes were read within the specified timeout

pub fn flush(&self) -> Result<(), EspError>

๐Deprecated since 0.41.3: Use UartRxDriver::clear instead

Clears the receive buffer.

pub fn clear(&self) -> Result<(), EspError>

pub fn port(&self) -> uart_port_t

pub fn count(&self) -> Result<usize, EspError>

Get count of remaining bytes in the receive ring buffer