esp_idf_hal::usb_serial
Struct UsbSerialDriver

pub struct UsbSerialDriver<'d>(/* private fields */);

USB-SERIAL driver

Implementations

impl<'d> UsbSerialDriver<'d>

pub fn new( usb_serial: impl Peripheral<P = USB_SERIAL> + 'd, _usb_d_min: impl Peripheral<P = UsbDMinGpio>, _usb_d_plus: impl Peripheral<P = UsbDPlusGpio>, config: &Config, ) -> Result<Self, EspError>

Create a new USB Serial driver

Arguments

usb_serial: The USB Serial peripheral

config: The driver configuration

usb_d_min: The USB D- GPIO pin

usb_d_plus: The USB D+ GPIO pin

pub fn is_connected(&self) -> bool

Check if the USB Serial is connected

pub fn read( &mut self, buf: &mut [u8], timeout: TickType_t, ) -> Result<usize, EspError>

Read bytes into a slice

Arguments

buf: The buffer to read into

timeout: The timeout in ticks

Returns

The number of bytes read or an error if the operation failed or the timeout was reached

pub fn write( &mut self, bytes: &[u8], timeout: TickType_t, ) -> Result<usize, EspError>

Write bytes from a slice

Arguments

bytes: The bytes to write

timeout: The timeout in ticks

Returns

The number of bytes written or an error if the operation failed or the timeout was reached