Crate esp_idf_hal

Modules

adc

can

CAN bus peripheral control.

cpu

delay

Delay providers.

gpio

GPIO and pin configuration

i2c

i2s

Driver for the Inter-IC Sound (I2S) peripheral(s).

interrupt

io

Error types

ledc

LED Control peripheral (which also creates PWM signals for other purposes)

modem

peripheral

peripherals

prelude

The prelude.

reset

Reset reasons

riscv_ulp_hal

rmt

Remote Control (RMT) module driver.

rom

ESP ROM libraries

spi

SPI peripheral control

sys

task

timer

uart

Controls UART peripherals (UART0, UART1, UART2). Notice that UART0 is typically already used for loading firmware and logging. Therefore use UART1 and UART2 in your application. Any pin can be used for rx and tx.

units

Units of measurement implementation for times and frequencies.

Macros

into_ref