esp_idf_hal::delay
Struct Delay

pub struct Delay(/* private fields */);

A delay provider that uses Ets for delays below a certain threshold and FreeRtos for delays equal or above the threshold.

Implementations

impl Delay

pub const fn new_default() -> Self

Create a Delay with a default threshold of 1 ms.

pub const fn new(threshold_us: u32) -> Self

Create a Delay with a threshold of the specified amount of microseconds.

pub fn delay_us(&self, us: u32)

Pauses execution for at minimum us microseconds. The delay can be longer due to rounding and/or runtime effects.

pub fn delay_ms(&self, ms: u32)

Pauses execution for at minimum ms milliseconds. The delay can be longer due to rounding and/or runtime effects.