esp_idf_hal
Module gpio

GPIO and pin configuration

Structs

AnyIOPin

Generic Gpio input-output pin

AnyInputPin

Generic Gpio input pin

AnyOutputPin

Generic Gpio output pin

Disabled

Gpio0

Gpio1

Gpio2

Gpio3

Gpio4

Gpio5

Gpio6

Gpio7

Gpio8

Gpio9

Gpio10

Gpio11

Gpio12

Gpio13

Gpio14

Gpio15

Gpio16

Gpio17

Gpio18

Gpio19

Gpio20

Gpio21

GpioError

Input

InputOutput

Output

PinDriver

A driver for a GPIO pin.

Pins

Enums

DriveStrength

Drive strength (values are approximates)

InterruptType

Interrupt types

Level

Digital input or output level.

Pull

Traits

ADCPin

A marker trait designating a pin which is capable of operating as an ADC pin

GPIOMode

IOPin

A marker trait designating a pin which is capable of operating as an input and output pin

InputMode

InputPin

A marker trait designating a pin which is capable of operating as an input pin

OutputMode

OutputPin

A marker trait designating a pin which is capable of operating as an output pin

Pin

A trait implemented by every pin instance

RTCPin

A marker trait designating a pin which is capable of operating as an RTC pin

Functions

enable_isr_service

init_isr_alloc_flags