Module esp_idf_hal::gpio

GPIO and pin configuration

Structs

Disabled

Gpio0

Gpio1

Gpio2

Gpio3

Gpio4

Gpio5

Gpio6

Gpio7

Gpio8

Gpio9

Gpio10

Gpio11

Gpio12

Gpio13

Gpio14

Gpio15

Gpio16

Gpio17

Gpio18

Gpio19

Gpio20

Gpio21

GpioPin

Generic $GpioX pin

Input

InputOutput

Output

Pins

SubscribedInput

Unknown

Enums

DriveStrength

Drive strength (values are approximates)

InterruptType

Interrupt types

Traits

ADCPin

A marker trait designating a pin which is capable of operating as an ADC pin, even if its current mode might be a different one

InputPin

A marker trait designating a pin which is capable of operating as an input pin, even if its current mode might be a different one

OutputPin

A marker trait designating a pin which is capable of operating as an output pin, even if its current mode might be a different one

Pin

A trait implemented by every pin insance

Pull

Functions available on pins with pull up/down resistors

RTCPin

SubscribedPin