esp_idf_hal::gpio
Struct AnyIOPin

pub struct AnyIOPin { /* private fields */ }

Generic Gpio input-output pin

Implementations

impl AnyIOPin

pub unsafe fn new(pin: i32) -> Self

Safety

Care should be taken not to instantiate this Pin, if it is already instantiated and used elsewhere, or if it is not set already in the mode of operation which is being instantiated

pub const fn none() -> Option<Self>

Creates an Option<AnyIOPin>::None for pins that are optional in APIs.