esp_idf_hal::gpio
Struct Gpio21

pub struct Gpio21(/* private fields */);

Implementations

impl Gpio21

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate this peripheral instance, if it is already instantiated and used elsewhere