Module esp_idf_hal::i2s

Driver for the Inter-IC Sound (I2S) peripheral(s).

Modules

config

I2S configuration

Structs

I2S0

I2sBiDir

Concrete implementation of both I2sRxSupported and I2sTxSupported for use in clients.

I2sDriver

Inter-IC Sound (I2S) driver.

I2sDriverRef

Reference for I2S driver

I2sRx

Concrete implementation of I2sRxSupported for use in clients.

I2sTx

Concrete implementation of I2sTxSupported for use in clients.

Traits

I2s

Trait implemented by I2S peripherals to obtain their port number.

I2sPort

I2sRxSupported

Marker trait indicating that a driver supports receiving data via the I2sRx trait.

I2sTxSupported

Marker trait indicating that a driver supports transmitting data via the I2sTx trait.

Type Aliases

I2sConfig

I2S channel base configuration.