esp_idf_hal::i2s

Module config

I2S configuration

Structs

Config

I2S common channel configuration.

HighPassFilter

PDM RX High Pass Filter

PdmRxClkConfig

Pulse density modulation (PDM) mode receive clock configuration for the I2S peripheral.

PdmRxConfig

Pulse density modulation (PDM) mode receive configuration for the I2S peripheral.

PdmRxGpioConfig

PDM mode GPIO (general purpose input/output) receive configuration.

PdmRxSlotConfig

PDM mode channel receive slot configuration.

PdmTxClkConfig

The I2s pulse density modulation (PDM) mode transmit clock configuration.

PdmTxConfig

The I2S pulse density modulation (PDM) mode transmit configuration for the I2S peripheral.

PdmTxGpioConfig

PDM mode GPIO (general purpose input/output) transmit configuration.

PdmTxSlotConfig

PDM mode channel transmit slot configuration.

StdClkConfig

Standard mode channel clock configuration.

StdConfig

Standard mode configuration for the I2S peripheral.

StdGpioConfig

Standard mode GPIO (general purpose input/output) configuration.

StdSlotConfig

Standard mode channel slot configuration.

TdmClkConfig

TDM mode channel clock configuration.

TdmConfig

Time-division multiplexing (TDM) mode configuration for the I2S peripheral.

TdmGpioConfig

TDM mode GPIO (general purpose input/output) configuration.

TdmSlotConfig

TDM mode slot configuration.

TdmSlotMask

Mask of TDM slots to enable.

Enums

ClockSource

I2S clock source.

DataBitWidth

Available data bit width in one slot.

MclkMultiple

The multiple of MCLK to the sample rate.

PdmDownsample

I2S pulse density modulation (PDM) downsampling mode.

PdmSignalScale

Pulse density modulation (PDM) transmit signal scaling mode.

PdmSlotMask

I2S slot selection in PDM mode.

PdmTxLineMode

I2S pulse density modulation (PDM) transmit line mode

Role

I2S channel operating role

SlotBitWidth

The total slot bit width in one slot.

SlotMode

I2S channel slot mode.

StdCommFormat

The communication format used by the v4 driver.

StdSlotMask

I2S slot selection in standard mode.

TdmSlot

An individual TDM slot.

Constants

DEFAULT_DMA_BUFFER_COUNT

The default number of DMA buffers to use.

DEFAULT_FRAMES_PER_DMA_BUFFER

The default number of frames per DMA buffer.

SOC_I2S_PDM_MAX_TX_LINES

The maximum number of data input pins that can be used in PDM mode.

TDM_AUTO_SLOT_NUM

Automatic total number of slots, equivalent to the maximum active slot number.

TDM_AUTO_WS_WIDTH

Automatic word-select signal width, equivalent to half the width of a frame.

TDM_BCLK_DIV_MIN

The minimum division from MCLK to BCLK.