esp_idf_hal::i2s
Struct I2sDriver

pub struct I2sDriver<'d, Dir> { /* private fields */ }

Inter-IC Sound (I2S) driver.

Implementations

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_pdm_tx<I2S: I2s>( _i2s: impl Peripheral<P = I2S> + 'd, tx_cfg: &PdmTxConfig, clk: impl Peripheral<P = impl OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, dout2: Option<impl Peripheral<P = impl OutputPin> + 'd>, ) -> Result<Self, EspError>

Create a new pulse density modulation (PDM) mode driver for the given I2S peripheral with only the transmit channel open.

impl<'d> I2sDriver<'d, I2sBiDir>

pub fn new_std_bidir<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with both the receive and transmit channels open.

impl<'d> I2sDriver<'d, I2sRx>

pub fn new_std_rx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with only the receive channel open.

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_std_tx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &StdConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new standard mode driver for the given I2S peripheral with only the transmit channel open.

impl<'d> I2sDriver<'d, I2sBiDir>

pub fn new_tdm_bidir<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with both the receive and transmit channels open.

impl<'d> I2sDriver<'d, I2sRx>

pub fn new_tdm_rx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, din: impl Peripheral<P = impl InputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with only the receive channel open.

impl<'d> I2sDriver<'d, I2sTx>

pub fn new_tdm_tx<I2S: I2s>( i2s: impl Peripheral<P = I2S> + 'd, config: &TdmConfig, bclk: impl Peripheral<P = impl InputPin + OutputPin> + 'd, dout: impl Peripheral<P = impl OutputPin> + 'd, mclk: Option<impl Peripheral<P = impl InputPin + OutputPin> + 'd>, ws: impl Peripheral<P = impl InputPin + OutputPin> + 'd, ) -> Result<Self, EspError>

Create a new TDM mode driver for the given I2S peripheral with only the transmit channel open.

impl<Dir> I2sDriver<'_, Dir>

pub fn as_ref(&mut self) -> I2sDriverRef<'_, Dir>

Borrow the I2S driver by a reference

impl<Dir> I2sDriver<'_, Dir>
where Dir: I2sRxSupported,

Functions for receive channels.

pub fn rx_enable(&mut self) -> Result<(), EspError>

Enable the I2S receive channel.

Note

This can only be called when the channel is in the READY state: initialized but not yet started from a driver constructor, or disabled from the RUNNING state via [rx_enable()][I2sRxChannel::rx_disable]. The channel will enter the RUNNING state if it is enabled successfully.

Enabling the channel will start I2S communications on the hardware. BCLK and WS signals will be generated if this is a controller. MCLK will be generated once initialization is finished.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the READY state.

pub fn rx_disable(&mut self) -> Result<(), EspError>

Disable the I2S receive channel.

Note

This can only be called when the channel is in the RUNNING state: the channel has been previously enabled via a call to [rx_enable()][I2sRxChannel::rx_enable]. The channel will enter the READY state if it is disabled successfully.

Disabling the channel will stop I2S communications on the hardware. BCLK and WS signals will stop being generated if this is a controller. MCLK will continue to be generated.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the RUNNING state.

pub async fn read_async(&mut self, buffer: &mut [u8]) -> Result<usize, EspError>

Read data from the channel asynchronously.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

pub fn read( &mut self, buffer: &mut [u8], timeout: TickType_t, ) -> Result<usize, EspError>

Read data from the channel.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

pub async fn read_uninit_async( &mut self, buffer: &mut [MaybeUninit<u8>], ) -> Result<usize, EspError>

Read data from the channel into an uninitalized buffer asynchronously.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

Safety

Upon a successful return with Ok(n_read), buffer[..n_read] will be initialized.

pub fn read_uninit( &mut self, buffer: &mut [MaybeUninit<u8>], timeout: TickType_t, ) -> Result<usize, EspError>

Read data from the channel into an uninitalized buffer.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes read, or an EspError if an error occurred.

Safety

Upon a successful return with Ok(n_read), buffer[..n_read] will be initialized.

impl<Dir> I2sDriver<'_, Dir>
where Dir: I2sTxSupported,

Functions for transmit channels.

pub fn tx_enable(&mut self) -> Result<(), EspError>

Enable the I2S transmit channel.

Note

This can only be called when the channel is in the READY state: initialized but not yet started from a driver constructor, or disabled from the RUNNING state via [tx_disable()][I2sTxChannel::tx_disable]. The channel will enter the RUNNING state if it is enabled successfully.

Enabling the channel will start I2S communications on the hardware. BCLK and WS signals will be generated if this is a controller. MCLK will be generated once initialization is finished.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the READY state.

pub fn tx_disable(&mut self) -> Result<(), EspError>

Disable the I2S transmit channel.

Note

This can only be called when the channel is in the RUNNING state: the channel has been previously enabled via a call to [tx_enable()][I2sTxChannel::tx_enable]. The channel will enter the READY state if it is disabled successfully.

Disabling the channel will stop I2S communications on the hardware. BCLK and WS signals will stop being generated if this is a controller. MCLK will continue to be generated.

Errors

This will return an EspError with ESP_ERR_INVALID_STATE if the channel is not in the RUNNING state.

pub fn preload_data(&mut self, data: &[u8]) -> Result<usize, EspError>

Preload data into the transmit channel DMA buffer.

This may be called only when the channel is in the READY state: initialized but not yet started.

This is used to preload data into the DMA buffer so that valid data can be transmitted immediately after the channel is enabled via [tx_enable()][I2sTxChannel::tx_enable]. If this function is not called before enabling the channel, empty data will be transmitted.

This function can be called multiple times before enabling the channel. Additional calls will concatenate the data to the end of the buffer until the buffer is full.

Returns

This returns the number of bytes that have been loaded into the buffer. If this is less than the length of the data provided, the buffer is full and no more data can be loaded.

pub async fn write_async(&mut self, data: &[u8]) -> Result<usize, EspError>

Write data to the channel asynchronously.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes sent. This may be less than the length of the data provided.

pub async fn write_all_async(&mut self, data: &[u8]) -> Result<(), EspError>

Write all data to the channel asynchronously.

This may be called only when the channel is in the RUNNING state.

pub fn write( &mut self, data: &[u8], timeout: TickType_t, ) -> Result<usize, EspError>

Write data to the channel.

This may be called only when the channel is in the RUNNING state.

Returns

This returns the number of bytes sent. This may be less than the length of the data provided.

pub fn write_all( &mut self, data: &[u8], timeout: TickType_t, ) -> Result<(), EspError>

Write all data to the channel.

This may be called only when the channel is in the RUNNING state.

impl I2sDriver<'_, I2sBiDir>

pub fn split(&mut self) -> (I2sDriverRef<'_, I2sRx>, I2sDriverRef<'_, I2sTx>)

Split the bidirectional I2S driver into two parts (Rx, Tx)

Safety

It is safe to use the two parts separately

esp-idf guarantees thread safety

esp-idf-hal guarantees asynchronous safety