esp_idf_hal
Module interrupt

Modules

asynch

Structs

IsrCriticalSection

IsrCriticalSectionGuard

Enums

InterruptType

Interrupt allocation flags. These flags can be used to specify which interrupt qualities the code calling esp_intr_alloc* needs.

Functions

active

Returns true if the currently active core is executing an ISR request

free

Executes closure f in an interrupt-free context

set_isr_yielder⚠

Safety

with_isr_yield_signal

Type Aliases

IntrFlags

For backwards compatibility