Module esp_idf_hal::io

Error types

Structs

EspIOError

Enums

ErrorKind

Possible kinds of errors.

ReadExactError

Error returned by Read::read_exact

SeekFrom

Enumeration of possible methods to seek within an I/O object.

SliceWriteError

Errors that could be returned by Write on &mut [u8].

WriteFmtError

Error returned by Write::write_fmt

Traits

BufRead

Blocking buffered reader.

Error

Error trait.

ErrorType

Base trait for all IO traits, defining the error type.

Read

Blocking reader.

ReadReady

Get whether a reader is ready.

Seek

Blocking seek within streams.

Write

Blocking writer.

WriteReady

Get whether a writer is ready.