Module esp_idf_hal::ledc

LED Control peripheral (which also creates PWM signals for other purposes)

Interface to the LED Control (LEDC) peripheral

This is an initial implementation supporting the generation of PWM signals but no chrome and spoilers like fading.

Examples

Create a 25 kHz PWM signal with 75 % duty cycle on GPIO 1

use embedded_hal::pwm::blocking::PwmPin;
use esp_idf_hal::ledc::{config::TimerConfig, Channel, Timer};
use esp_idf_hal::peripherals::Peripherals;
use esp_idf_hal::prelude::*;

let peripherals = Peripherals::take().unwrap();
let config = TimerConfig::default().frequency(25.kHz().into());
let timer = Timer::new(peripherals.ledc.timer0, &config)?;
let channel = Channel::new(peripherals.ledc.channel0, &timer, peripherals.pins.gpio1)?;

let max_duty = channel.get_max_duty()?;
channel.set_duty(max_duty * 3 / 4);

See the examples/ folder of this repository for more.

Modules

config

Types for configuring the LED Control peripheral

Structs

CHANNEL0

CHANNEL1

CHANNEL2

CHANNEL3

CHANNEL4

CHANNEL5

Channel

LED Control output channel abstraction

Peripheral

The LED Control device peripheral

TIMER0

TIMER1

TIMER2

TIMER3

Timer

LED Control timer abstraction

Enums

Resolution

Traits

HwChannel

LED Control peripheral output channel

HwTimer

LED Control peripheral timer