esp_idf_hal::peripherals

Struct Peripherals

pub struct Peripherals {

    pub pins: Pins,
    pub uart0: UART0<'static>,
    pub uart1: UART1<'static>,
    pub i2c0: I2C0<'static>,
    pub i2s0: I2S0<'static>,
    pub spi1: SPI1<'static>,
    pub spi2: SPI2<'static>,
    pub adc1: ADC1<'static>,
    pub adc2: ADC2<'static>,
    pub can: CAN<'static>,
    pub ledc: LEDC,
    pub modem: Modem<'static>,
    pub temp_sensor: TempSensor<'static>,
    pub twdt: TWDT<'static>,
    pub usb_serial: USB_SERIAL<'static>,

}

Fields
pins: Pinsuart0: UART0<'static>uart1: UART1<'static>i2c0: I2C0<'static>i2s0: I2S0<'static>spi1: SPI1<'static>spi2: SPI2<'static>adc1: ADC1<'static>adc2: ADC2<'static>can: CAN<'static>ledc: LEDCmodem: Modem<'static>temp_sensor: TempSensor<'static>twdt: TWDT<'static>usb_serial: USB_SERIAL<'static>
Implementations

impl Peripherals

pub fn take() -> Result<Self, EspError>

pub unsafe fn steal() -> Self

Safety

Care should be taken not to instantiate the Peripherals structure, if it is already instantiated and used elsewhere