esp_idf_hal::peripherals
Struct Peripherals

pub struct Peripherals {

    pub pins: Pins,
    pub uart0: UART0,
    pub uart1: UART1,
    pub i2c0: I2C0,
    pub i2s0: I2S0,
    pub spi1: SPI1,
    pub spi2: SPI2,
    pub adc1: ADC1,
    pub adc2: ADC2,
    pub can: CAN,
    pub ledc: LEDC,
    pub rmt: RMT,
    pub modem: Modem,
    pub temp_sensor: TempSensor,
    pub timer00: TIMER00,
    pub timer10: TIMER10,
    pub twdt: TWDT,
    pub usb_serial: USB_SERIAL,

}

Fields
pins: Pinsuart0: UART0uart1: UART1i2c0: I2C0i2s0: I2S0spi1: SPI1spi2: SPI2adc1: ADC1adc2: ADC2can: CANledc: LEDCrmt: RMTmodem: Modemtemp_sensor: TempSensortimer00: TIMER00timer10: TIMER10twdt: TWDTusb_serial: USB_SERIAL
Implementations

impl Peripherals

pub fn take() -> Result<Self, EspError>

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate the Peripherals structure, if it is already instantiated and used elsewhere