Struct esp_idf_hal::peripherals::Peripherals

pub struct Peripherals {
    pub pins: Pins,
    pub uart0: UART0,
    pub uart1: UART1,
    pub i2c0: I2C0,
    pub spi1: SPI1,
    pub spi2: SPI2,
    pub adc1: ADC1,
    pub adc2: ADC2,
    pub can: CAN,
    pub ledc: Peripheral,
    pub rmt: Peripheral,
}

Fields
pins: Pinsuart0: UART0uart1: UART1i2c0: I2C0spi1: SPI1spi2: SPI2adc1: ADC1adc2: ADC2can: CANledc: Peripheralrmt: Peripheral
Implementations

impl Peripherals

pub fn take() -> Option<Self>

pub unsafe fn new() -> Self

Safety

Care should be taken not to instantiate the Peripherals structure, if it is already instantiated and used elsewhere