Struct esp_idf_hal::rmt::FixedLengthSignal

pub struct FixedLengthSignal<const N: usize>(_);

Stack based signal storage for an RMT signal.

Use this if you know the length of the pulses ahead of time and prefer to use the stack.

Internally RMT uses pairs of pulses as part of its data structure. This implementation you need to set a two Pulses for each index.

let p1 = Pulse::new(PinState::High, PulseTicks::new(10));
let p2 = Pulse::new(PinState::Low, PulseTicks::new(11));
let p3 = Pulse::new(PinState::High, PulseTicks::new(12));
let p4 = Pulse::new(PinState::Low, PulseTicks::new(13));

let mut s = FixedLengthSignal::new();
s.set(0, &(p1, p2));
s.set(1, &(p3, p4));

Implementations

impl<const N: usize> FixedLengthSignal<N>

pub fn new() -> Self

Creates a new array of size <N>, where the number of pulses is N * 2.

pub fn set(
&mut self,
index: usize,
pair: &(Pulse, Pulse)
) -> Result<(), EspError>

Set a pair of Pulses at a position in the array.