esp_idf_hal::sys::build_time
Macro build_time_utc

build_time_utc!() { /* proc-macro */ }

Build time in UTC.

It takes an optional strftime date and time format string as input, and returns a string literal. If the input is empty, it will return a string literal in RFC 3339 date and time format, e.g., "2021-05-29T06:55:50.418437046+00:00".

Example

use build_time::build_time_utc;

// Returns the UTC build timestamp in RFC3339 date and time format.
let build_time_rfc3339 = build_time_utc!();

// Returns the UTC build timestamp in the specified format.
let build_time_formatted = build_time_utc!("%Y-%m-%dT%H:%M:%S%.f%:z");