esp_idf_hal::sys
Function sdmmc_io_get_cis_data

pub unsafe extern "C" fn sdmmc_io_get_cis_data(
    card: *mut sdmmc_card_t,
    out_buffer: *mut u8,
    buffer_size: usize,
    inout_cis_size: *mut usize,
) -> i32

Get the data of CIS region of an SDIO card.

You may provide a buffer not sufficient to store all the CIS data. In this case, this function stores as much data into your buffer as possible. Also, this function will try to get and return the size required for you.

@param card pointer to card information structure previously initialized using sdmmc_card_init @param out_buffer Output buffer of the CIS data @param buffer_size Size of the buffer. @param inout_cis_size Mandatory, pointer to a size, input and output. - input: Limitation of maximum searching range, should be 0 or larger than buffer_size. The function searches for CIS_CODE_END until this range. Set to 0 to search infinitely. - output: The size required to store all the CIS data, if CIS_CODE_END is found.

@return - ESP_OK: on success - ESP_ERR_INVALID_RESPONSE: if the card does not (correctly) support CIS. - ESP_ERR_INVALID_SIZE: CIS_CODE_END found, but buffer_size is less than required size, which is stored in the inout_cis_size then. - ESP_ERR_NOT_FOUND: if the CIS_CODE_END not found. Increase input value of inout_cis_size or set it to 0, if you still want to search for the end; output value of inout_cis_size is invalid in this case. - and other error code return from sdmmc_io_read_bytes