Type Alias esp_idf_hal::sys::TaskStatus_t

pub type TaskStatus_t = xTASK_STATUS;

Used with the uxTaskGetSystemState() function to return the state of each task in the system.

Aliased Type

struct TaskStatus_t {
    pub xHandle: *mut tskTaskControlBlock,
    pub pcTaskName: *const i8,
    pub xTaskNumber: u32,
    pub eCurrentState: u32,
    pub uxCurrentPriority: u32,
    pub uxBasePriority: u32,
    pub ulRunTimeCounter: u32,
    pub pxStackBase: *mut u8,
    pub usStackHighWaterMark: u32,
}

Fields
xHandle: *mut tskTaskControlBlock

< The handle of the task to which the rest of the information in the structure relates.

pcTaskName: *const i8

< A pointer to the task’s name. This value will be invalid if the task was deleted since the structure was populated!

xTaskNumber: u32

< A number unique to the task.

eCurrentState: u32

< The state in which the task existed when the structure was populated.

uxCurrentPriority: u32

< The priority at which the task was running (may be inherited) when the structure was populated.

uxBasePriority: u32

< The priority to which the task will return if the task’s current priority has been inherited to avoid unbounded priority inversion when obtaining a mutex. Only valid if configUSE_MUTEXES is defined as 1 in FreeRTOSConfig.h.

ulRunTimeCounter: u32

< The total run time allocated to the task so far, as defined by the run time stats clock. See https://www.FreeRTOS.org/rtos-run-time-stats.html. Only valid when configGENERATE_RUN_TIME_STATS is defined as 1 in FreeRTOSConfig.h.

pxStackBase: *mut u8

< Points to the lowest address of the task’s stack area.

usStackHighWaterMark: u32

< The minimum amount of stack space that has remained for the task since the task was created. The closer this value is to zero the closer the task has come to overflowing its stack.