esp_idf_hal::sys
Type Alias esp_netif_inherent_config_t

pub type esp_netif_inherent_config_t = esp_netif_inherent_config;

@brief ESP-netif inherent config parameters

Aliased Type

struct esp_netif_inherent_config_t {
    pub flags: u32,
    pub mac: [u8; 6],
    pub ip_info: *const esp_netif_ip_info_t,
    pub get_ip_event: u32,
    pub lost_ip_event: u32,
    pub if_key: *const u8,
    pub if_desc: *const u8,
    pub route_prio: i32,
    pub bridge_info: *mut bridgeif_config,
}

Fields
flags: u32

< flags that define esp-netif behavior

mac: [u8; 6]

< initial mac address for this interface

ip_info: *const esp_netif_ip_info_t

< initial ip address for this interface

get_ip_event: u32

< event id to be raised when interface gets an IP

lost_ip_event: u32

< event id to be raised when interface losts its IP

if_key: *const u8

< string identifier of the interface

if_desc: *const u8

< textual description of the interface

route_prio: i32

< numeric priority of this interface to become a default routing if (if other netifs are up). A higher value of route_prio indicates a higher priority

bridge_info: *mut bridgeif_config

< LwIP bridge configuration