esp_idf_hal::sys
Type Alias touch_filter_config_t

pub type touch_filter_config_t = touch_filter_config;

Touch sensor filter configuration

Aliased Type

struct touch_filter_config_t {
    pub mode: u32,
    pub debounce_cnt: u32,
    pub noise_thr: u32,
    pub jitter_step: u32,
    pub smh_lvl: u32,
}

Fields
mode: u32

<Set filter mode. The input of the filter is the raw value of touch reading, and the output of the filter is involved in the judgment of the touch state.

debounce_cnt: u32

<Set debounce count, such as n. If the measured values continue to exceed the threshold for n+1 times, the touch sensor state changes. Range: 0 ~ 7

noise_thr: u32

<Noise threshold coefficient. Higher = More noise resistance. The actual noise should be less than (noise coefficient * touch threshold). Range: 0 ~ 3. The coefficient is 0: 4/8; 1: 3/8; 2: 2/8; 3: 1;

jitter_step: u32

<Set jitter filter step size. Range: 0 ~ 15

smh_lvl: u32

<Level of filter applied on the original data against large noise interference.