esp_idf_hal
Module task

Modules

asynch

embassy_sync

notification

queue

thread

watchdog

Example

Structs

CriticalSection

CriticalSectionGuard

Functions

block_on

Executes the supplied future on the current thread, thus blocking it until the future becomes ready.

create⚠

Creates a FreeRTOS task.

current

destroy⚠

Deletes a FreeRTOS task.

do_yield

get_idle_task

notify⚠

Safety

notify_and_yield⚠

Safety

wait_notification

yield_now

Yield from the current task once, allowing other tasks to run.