esp_idf_hal::timer
Type Alias TimerConfig

pub type TimerConfig = Config;

Aliased Type

struct TimerConfig {
    pub divider: u32,
    pub xtal: bool,
    pub auto_reload: bool,
}

Fields
divider: u32xtal: boolauto_reload: bool

Enable or disable counter reload function when alarm event occurs.

Enabling this makes the hardware automatically reset the counter to the value set by TimerDriver::set_counter when the alarm is fired. This allows creating timers that automatically fire at a given interval without the software having to do anything after the timer setup.