esp_idf_hal::uart
Type Alias UartConfig

pub type UartConfig = Config;

Aliased Type

struct UartConfig {

    pub mode: Mode,
    pub baudrate: Hertz,
    pub data_bits: DataBits,
    pub parity: Parity,
    pub stop_bits: StopBits,
    pub flow_control: FlowControl,
    pub flow_control_rts_threshold: u8,
    pub source_clock: SourceClock,
    pub intr_flags: EnumSet<InterruptType>,
    pub event_config: EventConfig,
    pub rx_fifo_size: usize,
    pub tx_fifo_size: usize,
    pub queue_size: usize,
    /* private fields */

}

Fields
mode: Modebaudrate: Hertzdata_bits: DataBitsparity: Paritystop_bits: StopBitsflow_control: FlowControlflow_control_rts_threshold: u8source_clock: SourceClockintr_flags: EnumSet<InterruptType>

Configures the flags to use for interrupt allocation, e.g. priority to use for the interrupt.

Note that you should not set Iram here, because it will be automatically set depending on the value of CONFIG_UART_ISR_IN_IRAM.

event_config: EventConfig

Configures the interrupts the driver should enable.

rx_fifo_size: usize

The size of the software rx buffer. Must be bigger than the hardware FIFO.

tx_fifo_size: usize

The size of the software tx buffer. Must be bigger than the hardware FIFO or 0 to disable transmit buffering (note that this will make write operations block until data has been sent out).

queue_size: usize

Number of events that should fit into the event queue. Specify 0 to prevent the creation of an event queue.